cranelift_codegen

Module binemit

Binary machine code emission.

The binemit module contains code for translating Cranelift’s intermediate representation into binary machine code.

Structs

CodeInfo

Container for information about a vector of compiled code and its supporting read-only data.

Enums

Reloc

Relocation kinds for every ISA

Type Aliases

Addend

Addend to add to the symbol value.

CodeOffset

Offset in bytes from the beginning of the function.