Trait cranelift_codegen::binemit::RelocSink

pub trait RelocSink {
    fn reloc_ebb(&mut self, _: CodeOffset, _: Reloc, _: CodeOffset);

    fn reloc_external(
        &mut self, 
        _: CodeOffset, 
        _: Reloc, 
        _: &ExternalName, 
        _: Addend
    );

    fn reloc_jt(&mut self, _: CodeOffset, _: Reloc, _: JumpTable);
}

A trait for receiving relocations for code that is emitted directly into memory.

Required methods

fn reloc_ebb(&mut self, _: CodeOffset, _: Reloc, _: CodeOffset)

Add a relocation referencing an EBB at the current offset.

fn reloc_external(
&mut self,
_: CodeOffset,
_: Reloc,
_: &ExternalName,
_: Addend
)

Add a relocation referencing an external symbol at the current offset.

fn reloc_jt(&mut self, _: CodeOffset, _: Reloc, _: JumpTable)

Add a relocation referencing a jump table.

Loading content...