Module cranelift_codegen::cursor

Cursor library.

This module defines cursor data types that can be used for inserting instructions.

Structs

EncCursor
Encoding cursor.

FuncCursor
Function cursor.

Enums

CursorPosition
The possible positions of a cursor.

Traits

Cursor
All cursor types implement the Cursor which provides common navigation operations.