Struct cranelift_codegen::cursor::FuncCursor

pub struct FuncCursor<'f> {
    pub func: &'f mut Function,
    /* private fields */
}

Function cursor.

A FuncCursor holds a mutable reference to a whole ir::Function while keeping a position too. The function can be re-borrowed by accessing the public cur.func member.

This cursor is for use before legalization. The inserted instructions are not given an encoding.

Fields
func: &'f mut Function

The referenced function.

Implementations

impl<'f> FuncCursor<'f>

pub fn new(func: &'f mut Function) -> Self

Create a new FuncCursor pointing nowhere.

pub fn use_srcloc(&mut self, inst: Inst)

Use the source location of inst for future instructions.

pub fn ins(&mut self) -> InsertBuilder<'_, &mut FuncCursor<'f>>

Create an instruction builder that inserts an instruction at the current position.