Module cranelift_codegen::data_value

This module gives users to instantiate values that Cranelift understands. These values are used, for example, during interpretation and for wrapping immediates.

Structs

DisplayDataValues
Helper structure for printing bracket-enclosed vectors of DataValues.

Enums

DataValue
Represent a data value. Where Value is an SSA reference, DataValue is the type + value that would be referred to by a Value.

DataValueCastFailure
Record failures to cast DataValue.

Functions

write_data_value_list
Helper function for displaying Vec<DataValue>.