Enum cranelift_codegen::data_value::DataValue

pub enum DataValue {

    B(bool),
    I8(i8),
    I16(i16),
    I32(i32),
    I64(i64),
    I128(i128),
    U8(u8),
    U16(u16),
    U32(u32),
    U64(u64),
    U128(u128),
    F32(Ieee32),
    F64(Ieee64),
    V128([u8; 16]),

}

Represent a data value. Where Value is an SSA reference, DataValue is the type + value that would be referred to by a Value.

Variants

B(bool)

Tuple Fields

0: bool

I8(i8)

Tuple Fields

0: i8

I16(i16)

Tuple Fields

0: i16

I32(i32)

Tuple Fields

0: i32

I64(i64)

Tuple Fields

0: i64

I128(i128)

Tuple Fields

0: i128

U8(u8)

Tuple Fields

0: u8

U16(u16)

Tuple Fields

0: u16

U32(u32)

Tuple Fields

0: u32

U64(u64)

Tuple Fields

0: u64

U128(u128)

Tuple Fields

0: u128

F32(Ieee32)

Tuple Fields

0: Ieee32

F64(Ieee64)

Tuple Fields

0: Ieee64

V128([u8; 16])

Tuple Fields

0: [u8; 16]

Implementations

impl DataValue

pub fn from_integer(
imm: i128,
ty: Type
) -> Result<DataValue, DataValueCastFailure>

Try to cast an immediate integer (a wrapped i64 on most Cranelift instructions) to the given Cranelift Type.

pub fn ty(&self) -> Type

Return the Cranelift IR Type for this DataValue.

pub fn is_vector(&self) -> bool

Return true if the value is a vector (i.e. DataValue::V128).

pub fn is_bool(&self) -> bool

Return true if the value is a bool (i.e. DataValue::B).

pub fn write_to_slice(&self, dst: &mut [u8])

Write a DataValue to a slice.

Panics:

Panics if the slice does not have enough space to accommodate the DataValue

pub fn read_from_slice(src: &[u8], ty: Type) -> Self

Read a DataValue from a slice using a given Type.

Panics:

Panics if the slice does not have enough space to accommodate the DataValue

pub unsafe fn write_value_to(&self, p: *mut u128)

Write a DataValue to a memory location.

pub unsafe fn read_value_from(p: *const u128, ty: Type) -> Self

Read a DataValue from a memory location using a given Type.