Function cranelift_codegen::incremental_cache::try_finish_recompile

pub fn try_finish_recompile(
    func: &Function, 
    bytes: &[u8]
) -> Result<CompiledCode, RecompileError>

Given a function that’s been precompiled and its entry in the caching storage, try to shortcut compilation of the given function.

Precondition: the bytes must have retrieved from a cache store entry which hash value is strictly the same as the Function’s computed hash retrieved from compute_cache_key.