Enum cranelift_codegen::ir::entities::AnyEntity

pub enum AnyEntity {

    Function,
    Block(Block),
    Inst(Inst),
    Value(Value),
    StackSlot(StackSlot),
    DynamicStackSlot(DynamicStackSlot),
    DynamicType(DynamicType),
    GlobalValue(GlobalValue),
    JumpTable(JumpTable),
    Constant(Constant),
    FuncRef(FuncRef),
    SigRef(SigRef),
    Heap(Heap),
    Table(Table),
    StackLimit,

}

An opaque reference to any of the entities defined in this module that can appear in CLIF IR.

Variants

Function

The whole function.

Block(Block)

a basic block.

Inst(Inst)

An instruction.

Value(Value)

An SSA value.

StackSlot(StackSlot)

A stack slot.

DynamicStackSlot(DynamicStackSlot)

A dynamic stack slot.

DynamicType(DynamicType)

A dynamic type

GlobalValue(GlobalValue)

A Global value.

JumpTable(JumpTable)

A jump table.

Constant(Constant)

A constant.

FuncRef(FuncRef)

An external function.

SigRef(SigRef)

A function call signature.

Heap(Heap)

A heap.

Table(Table)

A table.

StackLimit

A function’s stack limit