Enum cranelift_codegen::ir::LibCall

pub enum LibCall {

    Probestack,
    UdivI64,
    SdivI64,
    UremI64,
    SremI64,
    IshlI64,
    UshrI64,
    SshrI64,
    CeilF32,
    CeilF64,
    FloorF32,
    FloorF64,
    TruncF32,
    TruncF64,
    NearestF32,
    NearestF64,
    Memcpy,
    Memset,
    Memmove,
    Memcmp,
    ElfTlsGetAddr,

}

The name of a runtime library routine.

Runtime library calls are generated for Cranelift IR instructions that don’t have an equivalent ISA instruction or an easy macro expansion. A LibCall is used as a well-known name to refer to the runtime library routine. This way, Cranelift doesn’t have to know about the naming convention in the embedding VM’s runtime library.

This list is likely to grow over time.

Variants

Probestack

probe for stack overflow. These are emitted for functions which need when the enable_probestack setting is true.

UdivI64

udiv.i64

SdivI64

sdiv.i64

UremI64

urem.i64

SremI64

srem.i64

IshlI64

ishl.i64

UshrI64

ushr.i64

SshrI64

sshr.i64

CeilF32

ceil.f32

CeilF64

ceil.f64

FloorF32

floor.f32

FloorF64

floor.f64

TruncF32

trunc.f32

TruncF64

frunc.f64

NearestF32

nearest.f32

NearestF64

nearest.f64

Memcpy

libc.memcpy

Memset

libc.memset

Memmove

libc.memmove

Memcmp

libc.memcmp

ElfTlsGetAddr

Elf __tls_get_addr

Implementations

impl LibCall

pub fn for_inst(opcode: Opcode, ctrl_type: Type) -> Option<Self>

Get the well-known library call name to use as a replacement for an instruction with the given opcode and controlling type variable.

Returns None if no well-known library routine name exists for that instruction.

pub fn all_libcalls() -> &'static [LibCall]

Get a list of all known LibCall’s.