cranelift_codegen::ir

Module function

Intermediate representation of a function.

The Function struct defined in this module owns all of its basic blocks and instructions.

Structs

DisplayFunction

Wrapper type capable of displaying a Function.

DisplayFunctionSpec

Wrapper type capable of displaying a ’Function’s name and signature.

Function

Functions can be cloned, but it is not a very fast operation. The clone will have all the same entity numbers as the original.

FunctionParameters

Function parameters used when creating this function, and that will become applied after compilation to materialize the final CompiledCode.

FunctionStencil

Function fields needed when compiling a function.

VersionMarker

A version marker used to ensure that serialized clif ir is never deserialized with a different version of Cranelift.