Struct cranelift_codegen::ir::function::Function

pub struct Function {
    pub name: ExternalName,
    pub signature: Signature,
    pub stack_slots: StackSlots,
    pub global_values: PrimaryMap<GlobalValue, GlobalValueData>,
    pub heaps: PrimaryMap<Heap, HeapData>,
    pub tables: PrimaryMap<Table, TableData>,
    pub jump_tables: JumpTables,
    pub dfg: DataFlowGraph,
    pub layout: Layout,
    pub encodings: InstEncodings,
    pub locations: ValueLocations,
    pub offsets: EbbOffsets,
    pub jt_offsets: JumpTableOffsets,
    pub srclocs: SourceLocs,
}

A function.

Functions can be cloned, but it is not a very fast operation. The clone will have all the same entity numbers as the original.

Fields
name: ExternalName

Name of this function. Mostly used by .clif files.

signature: Signature

Signature of this function.

stack_slots: StackSlots

Stack slots allocated in this function.

global_values: PrimaryMap<GlobalValue, GlobalValueData>

Global values referenced.

heaps: PrimaryMap<Heap, HeapData>

Heaps referenced.

tables: PrimaryMap<Table, TableData>

Tables referenced.

jump_tables: JumpTables

Jump tables used in this function.

dfg: DataFlowGraph

Data flow graph containing the primary definition of all instructions, EBBs and values.

layout: Layout

Layout of EBBs and instructions in the function body.

encodings: InstEncodings

Encoding recipe and bits for the legal instructions. Illegal instructions have the Encoding::default() value.

locations: ValueLocations

Location assigned to every value.

offsets: EbbOffsets

Code offsets of the EBB headers.

This information is only transiently available after the binemit::relax_branches function computes it, and it can easily be recomputed by calling that function. It is not included in the textual IR format.

jt_offsets: JumpTableOffsets

Code offsets of Jump Table headers.

srclocs: SourceLocs

Source locations.

Track the original source location for each instruction. The source locations are not interpreted by Cranelift, only preserved.

Methods

impl Function

pub fn with_name_signature(name: ExternalName, sig: Signature) -> Self

Create a function with the given name and signature.

pub fn clear(&mut self)

Clear all data structures in this function.

pub fn new() -> Self

Create a new empty, anonymous function with a Fast calling convention.

pub fn create_jump_table(&mut self, data: JumpTableData) -> JumpTable

Creates a jump table in the function, to be used by br_table instructions.

pub fn create_stack_slot(&mut self, data: StackSlotData) -> StackSlot

Creates a stack slot in the function, to be used by stack_load, stack_store and stack_addr instructions.

pub fn import_signature(&mut self, signature: Signature) -> SigRef

Adds a signature which can later be used to declare an external function import.

pub fn import_function(&mut self, data: ExtFuncData) -> FuncRef

Declare an external function import.

pub fn create_global_value(&mut self, data: GlobalValueData) -> GlobalValue

Declares a global value accessible to the function.

pub fn create_heap(&mut self, data: HeapData) -> Heap

Declares a heap accessible to the function.

pub fn create_table(&mut self, data: TableData) -> Table

Declares a table accessible to the function.

pub fn display<'a, I: Into<Option<&'a dyn TargetIsa>>>(
&'a self,
isa: I
) -> DisplayFunction<'a>

Return an object that can display this function with correct ISA-specific annotations.

pub fn special_param(&self, purpose: ArgumentPurpose) -> Option<Value>

Find a presumed unique special-purpose function parameter value.

Returns the value of the last purpose parameter, or None if no such parameter exists.

Important traits for InstOffsetIter<'a>
impl<'a> Iterator for InstOffsetIter<'a> type Item = (CodeOffset, Inst, CodeOffset);

pub fn inst_offsets<'a>(
&'a self,
ebb: Ebb,
encinfo: &EncInfo
) -> InstOffsetIter<'a>

Get an iterator over the instructions in ebb, including offsets and encoded instruction sizes.

The iterator returns (offset, inst, size) tuples, where offset if the offset in bytes from the beginning of the function to the instruction, and size is the size of the instruction in bytes, or 0 for unencoded instructions.

This function can only be used after the code layout has been computed by the binemit::relax_branches() function.

pub fn update_encoding(
&mut self,
inst: Inst,
isa: &dyn TargetIsa
) -> Result<(), Legalize>

Wrapper around encode which assigns inst the resulting encoding.

pub fn encode(
&self,
inst: Inst,
isa: &dyn TargetIsa
) -> Result<Encoding, Legalize>

Wrapper around TargetIsa::encode for encoding an existing instruction in the Function.