Struct cranelift_codegen::ir::immediates::Ieee64

#[repr(C)]
pub struct Ieee64(_);

An IEEE binary64 immediate floating point value, represented as a u64 containing the bit pattern.

We specifically avoid using a f64 here since some architectures may silently alter floats. See: https://github.com/bytecodealliance/wasmtime/pull/2251#discussion_r498508646

The PartialEq and Hash implementations are over the underlying bit pattern, but PartialOrd respects IEEE754 semantics.

All bit patterns are allowed.

Implementations

impl Ieee64

pub fn with_bits(x: u64) -> Self

Create a new Ieee64 containing the bits of x.

pub fn pow2<I: Into<i64>>(n: I) -> Self

Create an Ieee64 number representing 2.0^n.

pub fn fcvt_to_sint_negative_overflow<I: Into<i64>>(n: I) -> Self

Create an Ieee64 number representing the greatest negative value not convertable from f64 to a signed integer with width n.

pub fn neg(self) -> Self

Return self negated.

pub fn with_float(x: f64) -> Self

Create a new Ieee64 representing the number x.

pub fn bits(self) -> u64

Get the bitwise representation.

pub fn is_nan(&self) -> bool

Check if the value is a NaN. For Ieee64, this means checking that the 11 exponent bits are all set.

pub fn as_f64(self) -> f64

Converts Self to a rust f64

pub fn sqrt(self) -> Self

Returns the square root of self.

pub fn abs(self) -> Self

Computes the absolute value of self.

pub fn copysign(self, sign: Self) -> Self

Returns a number composed of the magnitude of self and the sign of sign.

pub fn is_negative(&self) -> bool

Returns true if self has a negative sign, including -0.0, NaNs with negative sign bit and negative infinity.

pub fn is_zero(&self) -> bool

Returns true if self is positive or negative zero

pub fn ceil(self) -> Self

Returns the smallest integer greater than or equal to self.

pub fn floor(self) -> Self

Returns the largest integer less than or equal to self.

pub fn trunc(self) -> Self

Returns the integer part of self. This means that non-integer numbers are always truncated towards zero.

pub fn round_ties_even(self) -> Self

Returns the nearest integer to self. Rounds half-way cases to the number with an even least significant digit.