cranelift_codegen::ir::instructions

Trait InstructionMapper

pub trait InstructionMapper {
    // Required methods
    fn map_value(&mut self, value: Value) -> Value;
    fn map_value_list(&mut self, value_list: ValueList) -> ValueList;
    fn map_global_value(&mut self, global_value: GlobalValue) -> GlobalValue;
    fn map_jump_table(&mut self, jump_table: JumpTable) -> JumpTable;
    fn map_exception_table(
        &mut self,
        exception_table: ExceptionTable,
    ) -> ExceptionTable;
    fn map_block_call(&mut self, block_call: BlockCall) -> BlockCall;
    fn map_func_ref(&mut self, func_ref: FuncRef) -> FuncRef;
    fn map_sig_ref(&mut self, sig_ref: SigRef) -> SigRef;
    fn map_stack_slot(&mut self, stack_slot: StackSlot) -> StackSlot;
    fn map_dynamic_stack_slot(
        &mut self,
        dynamic_stack_slot: DynamicStackSlot,
    ) -> DynamicStackSlot;
    fn map_constant(&mut self, constant: Constant) -> Constant;
    fn map_immediate(&mut self, immediate: Immediate) -> Immediate;
}

A trait to map some functions over each of the entities within an instruction, when paired with InstructionData::map.

Required Methods

fn map_value(&mut self, value: Value) -> Value

Map a function over a Value.

fn map_value_list(&mut self, value_list: ValueList) -> ValueList

Map a function over a ValueList.

fn map_global_value(&mut self, global_value: GlobalValue) -> GlobalValue

Map a function over a GlobalValue.

fn map_jump_table(&mut self, jump_table: JumpTable) -> JumpTable

Map a function over a JumpTable.

fn map_exception_table( &mut self, exception_table: ExceptionTable, ) -> ExceptionTable

Map a function over an ExceptionTable.

fn map_block_call(&mut self, block_call: BlockCall) -> BlockCall

Map a function over a BlockCall.

fn map_func_ref(&mut self, func_ref: FuncRef) -> FuncRef

Map a function over a FuncRef.

fn map_sig_ref(&mut self, sig_ref: SigRef) -> SigRef

Map a function over a SigRef.

fn map_stack_slot(&mut self, stack_slot: StackSlot) -> StackSlot

Map a function over a StackSlot.

fn map_dynamic_stack_slot( &mut self, dynamic_stack_slot: DynamicStackSlot, ) -> DynamicStackSlot

Map a function over a DynamicStackSlot.

fn map_constant(&mut self, constant: Constant) -> Constant

Map a function over a Constant.

fn map_immediate(&mut self, immediate: Immediate) -> Immediate

Map a function over an Immediate.