Struct cranelift_codegen::ir::stackslot::StackSlots

pub struct StackSlots {
    pub frame_size: Option<StackSize>,
    // some fields omitted
}

Stack frame manager.

Keep track of all the stack slots used by a function.

Fields
frame_size: Option<StackSize>

The total size of the stack frame.

This is the distance from the stack pointer in the current function to the stack pointer in the calling function, so it includes a pushed return address as well as space for outgoing call arguments.

This is computed by the layout() method.

Methods

impl StackSlots

Stack slot manager functions that behave mostly like an entity map.

pub fn new() -> Self

Create an empty stack slot manager.

pub fn clear(&mut self)

Clear out everything.

pub fn push(&mut self, data: StackSlotData) -> StackSlot

Allocate a new stack slot.

This function should be primarily used by the text format parser. There are more convenient functions for creating specific kinds of stack slots below.

pub fn is_valid(&self, ss: StackSlot) -> bool

Check if ss is a valid stack slot reference.

pub fn iter(&self) -> Iter<StackSlot, StackSlotData>

Get an iterator over all the stack slot keys.

pub fn iter_mut(&mut self) -> IterMut<StackSlot, StackSlotData>

Get an iterator over all the stack slot keys, mutable edition.

pub fn values(&self) -> Iter<StackSlotData>

Get an iterator over all the stack slot records.

pub fn values_mut(&mut self) -> IterMut<StackSlotData>

Get an iterator over all the stack slot records, mutable edition.

pub fn keys(&self) -> Keys<StackSlot>

Get an iterator over all the stack slot keys.

pub fn next_key(&self) -> StackSlot

Get a reference to the next stack slot that would be created by push().

This should just be used by the parser.

impl StackSlots

Higher-level stack frame manipulation functions.

pub fn make_spill_slot(&mut self, ty: Type) -> StackSlot

Create a new spill slot for spilling values of type ty.

pub fn make_incoming_arg(&mut self, ty: Type, offset: StackOffset) -> StackSlot

Create a stack slot representing an incoming function argument.

pub fn get_outgoing_arg(&mut self, ty: Type, offset: StackOffset) -> StackSlot

Get a stack slot representing an outgoing argument.

This may create a new stack slot, or reuse an existing outgoing stack slot with the requested offset and size.

The requested offset is relative to this function's stack pointer immediately before making the call.

pub fn get_emergency_slot(
&mut self,
ty: Type,
in_use: &[PackedOption<StackSlot>]
) -> StackSlot

Get an emergency spill slot that can be used to store a ty value.

This may allocate a new slot, or it may reuse an existing emergency spill slot, excluding any slots in the in_use list.