cranelift_codegen::ir

Struct ReplaceBuilder

pub struct ReplaceBuilder<'f> { /* private fields */ }

Instruction builder that replaces an existing instruction.

The inserted instruction will have the same Inst number as the old one.

If the old instruction still has result values attached, it is assumed that the new instruction produces the same number and types of results. The old result values are preserved. If the replacement instruction format does not support multiple results, the builder panics. It is a bug to leave result values dangling.

Implementations

impl<'f> ReplaceBuilder<'f>

pub fn new(dfg: &'f mut DataFlowGraph, inst: Inst) -> Self

Create a ReplaceBuilder that will overwrite inst.