Module cranelift_codegen::isa::aarch64::inst

This module defines aarch64-specific machine instruction types.

Re-exports

pub use self::imms::*;

pub use self::args::*;

Modules

args

AArch64 ISA definitions: instruction arguments.

imms

AArch64 ISA definitions: immediate constants.

Structs

CallIndInfo

Additional information for CallInd instructions, left out of line to lower the size of the Inst enum.

CallInfo

Additional information for (direct) Call instructions, left out of line to lower the size of the Inst enum.

JTSequenceInfo

Additional information for JTSequence instructions, left out of line to lower the size of the Inst enum.

Enums

ALUOp

Internal type ALUOp: defined at src/isa/aarch64/inst.isle line 951.

ALUOp3

Internal type ALUOp3: defined at src/isa/aarch64/inst.isle line 989.

AMode

Internal type AMode: defined at src/isa/aarch64/inst.isle line 1052.

APIKey

Internal type APIKey: defined at src/isa/aarch64/inst.isle line 1586.

AtomicRMWLoopOp

Internal type AtomicRMWLoopOp: defined at src/isa/aarch64/inst.isle line 1570.

AtomicRMWOp

Internal type AtomicRMWOp: defined at src/isa/aarch64/inst.isle line 1555.

BitOp

Internal type BitOp: defined at src/isa/aarch64/inst.isle line 1035.

BranchTargetType

Internal type BranchTargetType: defined at src/isa/aarch64/inst.isle line 1593.

FPUOp1

Internal type FPUOp1: defined at src/isa/aarch64/inst.isle line 1240.

FPUOp2

Internal type FPUOp2: defined at src/isa/aarch64/inst.isle line 1250.

FPUOp3

Internal type FPUOp3: defined at src/isa/aarch64/inst.isle line 1261.

FPUOpRI

A floating-point unit (FPU) operation with two args, a register and an immediate.

FPUOpRIMod

A floating-point unit (FPU) operation with two args, a register and an immediate that modifies its dest (so takes that input value as a separate virtual register).

FpuRoundMode

Internal type FpuRoundMode: defined at src/isa/aarch64/inst.isle line 1294.

FpuToIntOp

Internal type FpuToIntOp: defined at src/isa/aarch64/inst.isle line 1267.

Inst

Internal type MInst: defined at src/isa/aarch64/inst.isle line 2.

IntToFpuOp

Internal type IntToFpuOp: defined at src/isa/aarch64/inst.isle line 1280.

LabelUse

Different forms of label references for different instruction formats.

MoveWideOp

Internal type MoveWideOp: defined at src/isa/aarch64/inst.isle line 997.

VecALUModOp

Internal type VecALUModOp: defined at src/isa/aarch64/inst.isle line 1393.

VecALUOp

Internal type VecALUOp: defined at src/isa/aarch64/inst.isle line 1316.

VecExtendOp

Internal type VecExtendOp: defined at src/isa/aarch64/inst.isle line 1307.

VecLanesOp

Internal type VecLanesOp: defined at src/isa/aarch64/inst.isle line 1528.

VecMisc2

Internal type VecMisc2: defined at src/isa/aarch64/inst.isle line 1402.

VecPairOp

Internal type VecPairOp: defined at src/isa/aarch64/inst.isle line 1509.

VecRRLongOp

Internal type VecRRLongOp: defined at src/isa/aarch64/inst.isle line 1459.

VecRRNarrowOp

Internal type VecRRNarrowOp: defined at src/isa/aarch64/inst.isle line 1474.

VecRRPairLongOp

Internal type VecRRPairLongOp: defined at src/isa/aarch64/inst.isle line 1517.

VecRRRLongModOp

Internal type VecRRRLongModOp: defined at src/isa/aarch64/inst.isle line 1500.

VecRRRLongOp

Internal type VecRRRLongOp: defined at src/isa/aarch64/inst.isle line 1488.

VecShiftImmModOp

Internal type VecShiftImmModOp: defined at src/isa/aarch64/inst.isle line 1548.

VecShiftImmOp

Internal type VecShiftImmOp: defined at src/isa/aarch64/inst.isle line 1537.