Enum cranelift_codegen::isa::CallConv

pub enum CallConv {
    Fast,
    Cold,
    SystemV,
    WindowsFastcall,
    BaldrdashSystemV,
    BaldrdashWindows,
    Probestack,
}

Calling convention identifiers.

Variants

Fast

Best performance, not ABI-stable

Cold

Smallest caller code size, not ABI-stable

SystemV

System V-style convention used on many platforms

WindowsFastcall

Windows "fastcall" convention, also used for x64 and ARM

BaldrdashSystemV

SpiderMonkey WebAssembly convention on systems using natively SystemV

BaldrdashWindows

SpiderMonkey WebAssembly convention on Windows

Probestack

Specialized convention for the probestack function

Methods

impl CallConv

pub fn triple_default(triple: &Triple) -> Self

Return the default calling convention for the given target triple.

pub fn for_libcall(isa: &dyn TargetIsa) -> Self

Returns the calling convention used for libcalls for the given ISA.

pub fn extends_windows_fastcall(self) -> bool

Is the calling convention extending the Windows Fastcall ABI?

pub fn extends_baldrdash(self) -> bool

Is the calling convention extending the Baldrdash ABI?