Module cranelift_codegen::isa::registers

Data structures describing the registers in an ISA.

Structs

DisplayRegUnit
Temporary object that holds enough information to print a register unit.

RegBank
The register units in a target ISA are divided into disjoint register banks. Each bank covers a contiguous range of register units.

RegClassData
Data about a register class.

RegClassIndex
A small reference to a register class.

RegInfo
Information about the registers in an ISA.

Functions

regs_overlap
Test of two registers overlap.

Type Definitions

RegClass
A register class reference.

RegClassMask
A bit mask indexed by register classes.

RegUnit
Register units are the smallest units of register allocation.

RegUnitMask
A bit mask indexed by register units.