Struct cranelift_codegen::isa::registers::RegBank

pub struct RegBank {
    pub name: &'static str,
    pub first_unit: RegUnit,
    pub units: RegUnit,
    pub names: &'static [&'static str],
    pub prefix: &'static str,
    pub first_toprc: usize,
    pub num_toprcs: usize,
    pub pressure_tracking: bool,
}

The register units in a target ISA are divided into disjoint register banks. Each bank covers a contiguous range of register units.

The RegBank struct provides a static description of a register bank.

Fields
name: &'static str

The name of this register bank as defined in the ISA’s DSL definition.

first_unit: RegUnit

The first register unit in this bank.

units: RegUnit

The total number of register units in this bank.

names: &'static [&'static str]

Array of specially named register units. This array can be shorter than the number of units in the bank.

prefix: &'static str

Name prefix to use for those register units in the bank not covered by the names array. The remaining register units will be named this prefix followed by their decimal offset in the bank. So with a prefix r, registers will be named r8, r9, …

first_toprc: usize

Index of the first top-level register class in this bank.

num_toprcs: usize

Number of top-level register classes in this bank.

The top-level register classes in a bank are guaranteed to be numbered sequentially from first_toprc, and all top-level register classes across banks come before any sub-classes.

pressure_tracking: bool

Is register pressure tracking enabled for this bank?