cranelift_codegen::isa
Module x64

X86_64-bit Instruction Set Architecture.

Modules

args

Instruction operand sub-components (aka “parts”): definitions and printing.

encoding

Contains the encoding machinery for the various x64 instruction formats.

settings

x86 Settings.

Structs

EmitInfo

Constant state used during emissions of a sequence of instructions.

EmitState

State carried between emissions of a sequence of instructions.

Enums

Inst

Internal type MInst: defined at src/isa/x64/inst.isle line 7.

Functions

create_cie

Creates a new x86-64 common information entry (CIE).

emit_unwind_info

Emit unwind info for an x86 target.