Module cranelift_codegen::isa::x64::encoding::vex

Encodes VEX instructions. These instructions are those added by the Advanced Vector Extensions (AVX).

Structs

VexInstruction

Constructs a VEX-encoded instruction using a builder pattern. This approach makes it visually easier to transform something the manual’s syntax, VEX.128.66.0F 73 /7 ib to code: VexInstruction::new().length(...).prefix(...).map(...).w(true).opcode(0x1F).reg(...).rm(...).

Enums

VexVectorLength

The VEX format allows choosing a vector length in the L bit.