Trait cranelift_codegen::machinst::pretty_print::ShowWithRRU

pub trait ShowWithRRU {
    fn show_rru(&self, mb_rru: Option<&RealRegUniverse>) -> String;

    fn show_rru_sized(
        &self, 
        mb_rru: Option<&RealRegUniverse>, 
        _size: u8
    ) -> String { ... }
}

A trait for printing instruction bits and pieces, with the the ability to take a contextualising RealRegUniverse that is used to give proper names to registers.

Required methods

fn show_rru(&self, mb_rru: Option<&RealRegUniverse>) -> String

Return a string that shows the implementing object in context of the given RealRegUniverse, if provided.

Loading content...
Provided methods

fn show_rru_sized(&self, mb_rru: Option<&RealRegUniverse>, _size: u8) -> String

The same as |show_rru|, but with an optional hint giving a size in bytes. Its interpretation is object-dependent, and it is intended to pass around enough information to facilitate printing sub-parts of real registers correctly. Objects may ignore size hints that are irrelevant to them.

Loading content...