Trait cranelift_codegen::machinst::MachBackend

pub trait MachBackend {
    fn compile_function(
        &self, 
        func: &Function, 
        want_disasm: bool
    ) -> CodegenResult<MachCompileResult>;

    fn flags(&self) -> &Flags;

    fn triple(&self) -> Triple;

    fn name(&self) -> &'static str;

    fn reg_universe(&self) -> &RealRegUniverse;

    fn unsigned_add_overflow_condition(&self) -> IntCC;

    fn unsigned_sub_overflow_condition(&self) -> IntCC;
}

Top-level machine backend trait, which wraps all monomorphized code and allows a virtual call from the machine-independent Function::compile().

Required methods

fn compile_function(
&self,
func: &Function,
want_disasm: bool
) -> CodegenResult<MachCompileResult>

Compile the given function.

fn flags(&self) -> &Flags

Return flags for this backend.

fn triple(&self) -> Triple

Return triple for this backend.

fn name(&self) -> &'static str

Return name for this backend.

fn reg_universe(&self) -> &RealRegUniverse

Return the register universe for this backend.

fn unsigned_add_overflow_condition(&self) -> IntCC

Machine-specific condcode info needed by TargetIsa. Condition that will be true when an IaddIfcout overflows.

fn unsigned_sub_overflow_condition(&self) -> IntCC

Machine-specific condcode info needed by TargetIsa. Condition that will be true when an IsubIfcout overflows.

Loading content...