Struct cranelift_codegen::machinst::valueregs::ValueRegs

pub struct ValueRegs<R: Clone + Copy + Debug + PartialEq + Eq + InvalidSentinel> { /* fields omitted */ }

Location at which a Value is stored in register(s): the value is located in one or more registers, depending on its width. A value may be stored in more than one register if the machine has no registers wide enough otherwise: for example, on a 32-bit architecture, we may store I64 values in two registers, and I128 values in four.

By convention, the register parts are kept in machine-endian order here.

N.B.: we cap the capacity of this at four (when any 32-bit target is enabled) or two (otherwise), and we use special in-band sentinal Reg values (Reg::invalid()) to avoid the need to carry a separate length. This allows the struct to be Copy (no heap or drop overhead) and be only 16 or 8 bytes, which is important for compiler performance.

Implementations

impl<R: Clone + Copy + Debug + PartialEq + Eq + InvalidSentinel> ValueRegs<R>

pub fn invalid() -> Self

Create an invalid Value-in-Reg.

pub fn is_valid(self) -> bool

Is this Value-to-Reg mapping valid?

pub fn is_invalid(self) -> bool

Is this Value-to-Reg mapping invalid?

pub fn only_reg(self) -> Option<R>

Return the single register used for this value, if any.

pub fn regs(&self) -> &[R]

Return an iterator over the registers storing this value.

impl<R: Clone + Copy + Debug + PartialEq + Eq + InvalidSentinel> ValueRegs<R>

pub fn one(reg: R) -> Self

Create a Value-in-R location for a value stored in one register.

pub fn two(r1: R, r2: R) -> Self

Create a Value-in-R location for a value stored in two registers.

pub fn len(self) -> usize

Return the number of registers used.

pub fn map<NewR, F>(self, f: F) -> ValueRegs<NewR> where
NewR: Clone + Copy + Debug + PartialEq + Eq + InvalidSentinel,
F: Fn(R) -> NewR,

Map individual registers via a map function.