Enum cranelift_codegen::machinst::vcode::VCodeConstantData

pub enum VCodeConstantData {
    Pool(Constant, ConstantData),
    WellKnown(&'static [u8]),
    Generated(ConstantData),
}

Identify the different types of constant that can be inserted into VCodeConstants. Tracking these separately instead of as raw byte buffers allows us to avoid some duplication.

Variants

Pool(Constant, ConstantData)

A constant already present in the Cranelift IR ConstantPool.

Tuple Fields of Pool

0: Constant1: ConstantData

WellKnown(&'static [u8])

A reference to a well-known constant value that is statically encoded within the compiler.

Tuple Fields of WellKnown

0: &'static [u8]

Generated(ConstantData)

A constant value generated during lowering; the value may depend on the instruction context which makes it difficult to de-duplicate–if possible, use other variants.

Tuple Fields of Generated

0: ConstantData

Implementations

impl VCodeConstantData

pub fn as_slice(&self) -> &[u8]

Retrieve the constant data as a byte slice.

pub fn alignment(&self) -> u32

Calculate the alignment of the constant data.