Struct cranelift_codegen::machinst::vcode::VCode

pub struct VCode<I: VCodeInst> { /* fields omitted */ }

A function in “VCode” (virtualized-register code) form, after lowering. This is essentially a standard CFG of basic blocks, where each basic block consists of lowered instructions produced by the machine-specific backend.

Implementations

impl<I: VCodeInst> VCode<I>

pub fn flags(&self) -> &Flags

Returns the flags controlling this function’s compilation.

pub fn vreg_type(&self, vreg: VirtualReg) -> Type

Get the IR-level type of a VReg.

pub fn have_ref_values(&self) -> bool

Are there any reference-typed values at all among the vregs?

pub fn entry(&self) -> BlockIndex

Get the entry block.

pub fn num_blocks(&self) -> usize

Get the number of blocks. Block indices will be in the range 0 .. (self.num_blocks() - 1).

pub fn frame_size(&self) -> u32

Stack frame size for the full function’s body.

pub fn stack_args_size(&self) -> u32

Inbound stack-args size.

pub fn succs(&self, block: BlockIndex) -> &[BlockIx]

Get the successors for a block.

pub fn replace_insns_from_regalloc(&mut self, result: RegAllocResult<Self>)

Take the results of register allocation, with a sequence of instructions including spliced fill/reload/move instructions, and replace the VCode with them.

pub fn emit(&self) -> MachBuffer<I> where
I: MachInstEmit,

Emit the instructions to a MachBuffer, containing fixed-up code and external reloc/trap/etc. records ready for use.

pub fn unwind_info(&self) -> CodegenResult<Option<UnwindInfo<Reg>>>

Generates unwind info.

pub fn bindex_to_bb(&self, block: BlockIndex) -> Option<Block>

Get the IR block for a BlockIndex, if one exists.