Module cranelift_codegen::settings::detail

Implementation details for generated code.

This module holds definitions that need to be public so the can be instantiated by generated code in other modules.

Structs

Descriptor
A setting descriptor holds the information needed to generically set and print a setting.

Template
An instruction group template.

Enums

Detail
The different kind of settings along with descriptor bits that depend on the kind.