Trait cranelift_codegen::settings::Configurable

pub trait Configurable {
    // Required methods
    fn set(&mut self, name: &str, value: &str) -> SetResult<()>;
    fn enable(&mut self, name: &str) -> SetResult<()>;
}

A string-based configurator for settings groups.

The Configurable protocol allows settings to be modified by name before a finished Flags struct is created.

Required Methods

fn set(&mut self, name: &str, value: &str) -> SetResult<()>

Set the string value of any setting by name.

This can set any type of setting whether it is numeric, boolean, or enumerated.

fn enable(&mut self, name: &str) -> SetResult<()>

Enable a boolean setting or apply a preset.

If the identified setting isn’t a boolean or a preset, a BadType error is returned.