cranelift_codegen

Struct Reg

pub struct Reg(/* private fields */);

A register named in an instruction. This register can be a virtual register, a fixed physical register, or a named spillslot (after regalloc). It does not have any constraints applied to it: those can be added later in MachInst::get_operands() when the Regs are converted to Operands.

Implementations

impl Reg

pub const fn from_virtual_reg(vreg: VReg) -> Reg

Const constructor: create a new Reg from a regalloc2 VReg.

pub const fn from_real_reg(preg: PReg) -> Reg

Const constructor: create a new Reg from a regalloc2 PReg.

pub fn to_real_reg(self) -> Option<RealReg>

Get the physical register (RealReg), if this register is one.

pub fn to_virtual_reg(self) -> Option<VirtualReg>

Get the virtual (non-physical) register, if this register is one.

pub fn to_spillslot(self) -> Option<SpillSlot>

Get the spillslot, if this register is one.

pub fn class(self) -> RegClass

Get the class of this register.

pub fn is_real(self) -> bool

Is this a real (physical) reg?

pub fn is_virtual(self) -> bool

Is this a virtual reg?

pub fn is_spillslot(self) -> bool

Is this a spillslot?