cranelift_codegen::verifier

Struct VerifierErrors

pub struct VerifierErrors(pub Vec<VerifierError>);

List of verifier errors.

Tuple Fields
0: Vec<VerifierError>
Implementations

impl VerifierErrors

pub fn new() -> Self

Return a new VerifierErrors struct.

pub fn is_empty(&self) -> bool

Return whether no errors were reported.

pub fn has_error(&self) -> bool

Return whether one or more errors were reported.

pub fn as_result(&self) -> VerifierStepResult

Return a VerifierStepResult that is fatal if at least one error was reported, and non-fatal otherwise.

pub fn report(&mut self, error: impl Into<VerifierError>)

Report an error, adding it to the list of errors.

pub fn fatal(&mut self, error: impl Into<VerifierError>) -> VerifierStepResult

Report a fatal error and return Err.

pub fn nonfatal( &mut self, error: impl Into<VerifierError>, ) -> VerifierStepResult

Report a non-fatal error and return Ok.