Type Definition cranelift_codegen::verifier::VerifierStepResult

type VerifierStepResult<T> = Result<T, ()>;

Result of a step in the verification process.

Functions that return VerifierStepResult<()> should also take a mutable reference to VerifierErrors as argument in order to report errors.

Here, Ok represents a step that did not lead to a fatal error, meaning that the verification process may continue. However, other (non-fatal) errors might have been reported through the previously mentioned VerifierErrors argument.