Crate cranelift_codegen

Cranelift code generation library.

Re-exports

pub use crate::verifier::verify_function;

pub use crate::write::write_function;

pub use cranelift_bforest as bforest;

pub use cranelift_entity as entity;

Modules

binemit
Binary machine code emission.

cfg_printer
The CFGPrinter utility.

cursor
Cursor library.

dbg
Debug tracing helpers.

dominator_tree
A Dominator Tree represented as mappings of Ebbs to their immediate dominator.

flowgraph
A control flow graph represented as mappings of extended basic blocks to their predecessors and successors.

ir
Representation of Cranelift IR functions.

isa
Instruction Set Architectures.

loop_analysis
A loop analysis represented as mappings of loops to their header Ebb and parent in the loop tree.

packed_option
Compact representation of Option<T> for types with a reserved value.

print_errors
Utility routines for pretty-printing error messages.

settings
Shared settings module.

timing
Pass timing.

verifier
A verifier for ensuring that functions are well formed. It verifies:

write
Converting Cranelift IR to text.

Structs

Context
Persistent data structures and compilation pipeline.

ValueLocRange
Value location range.

Enums

CodegenError
A compilation error.

Constants

VERSION
Version number of this crate.

Functions

legalize_function
Legalize func for isa.

Type Definitions

CodegenResult
A convenient alias for a Result that uses CodegenError as the error type.

ValueLabelsRanges
Resulting map of Value labels and their ranges/locations.