Struct cranelift_codegen::binemit::RegDiversions

pub struct RegDiversions { /* fields omitted */ }

Keep track of diversions in a block.

Implementations

impl RegDiversions

pub fn new() -> Self

Create a new empty diversion tracker.

pub fn clear(&mut self)

Clear the content of the diversions, to reset the state of the compiler.

pub fn is_empty(&self) -> bool

Are there any diversions?

pub fn diversion(&self, value: Value) -> Option<&Diversion>

Get the current diversion of value, if any.

pub fn iter(&self) -> Iter<'_, Value, Diversion>

Get all current diversions.

pub fn get(&self, value: Value, locations: &ValueLocations) -> ValueLoc

Get the current location for value. Fall back to the assignment map for non-diverted values

pub fn reg(&self, value: Value, locations: &ValueLocations) -> RegUnit

Get the current register location for value, or panic if value isn’t in a register.

pub fn stack(&self, value: Value, locations: &ValueLocations) -> StackSlot

Get the current stack location for value, or panic if value isn’t in a stack slot.

pub fn regmove(&mut self, value: Value, from: RegUnit, to: RegUnit)

Record a register -> register move.

pub fn regspill(&mut self, value: Value, from: RegUnit, to: StackSlot)

Record a register -> stack move.

pub fn regfill(&mut self, value: Value, from: StackSlot, to: RegUnit)

Record a stack -> register move.

pub fn apply(&mut self, inst: &InstructionData)

Apply the effect of inst.

If inst is a regmove, regfill, or regspill instruction, update the diversions to match.

pub fn remove(&mut self, value: Value) -> Option<ValueLoc>

Drop any recorded move for value.

Returns the to location of the removed diversion.

pub fn at_block(&mut self, entry_diversions: &EntryRegDiversions, block: Block)

Resets the state of the current diversions to the recorded diversions at the entry of the given block. The recoded diversions is available after coloring on func.entry_diversions field.

pub fn save_for_block(
&mut self,
entry_diversions: &mut EntryRegDiversions,
target: Block
)

Copy the current state of the diversions, and save it for the entry of the block given as argument.

Note: This function can only be called once on a Block with a given entry_diversions argument, otherwise it would panic.

pub fn check_block_entry(
&self,
entry_diversions: &EntryRegDiversions,
target: Block
) -> bool

Check that the recorded entry for a given block matches what is recorded in the entry_diversions.

pub fn display<'a, R: Into<Option<&'a RegInfo>>>(
&'a self,
regs: R
) -> DisplayDiversions<'a>

Return an object that can display the diversions.