Struct cranelift_codegen::cursor::EncCursor

pub struct EncCursor<'f> {
    pub func: &'f mut Function,
    pub isa: &'f dyn TargetIsa,
    // some fields omitted
}

Encoding cursor.

An EncCursor can be used to insert instructions that are immediately assigned an encoding. The cursor holds a mutable reference to the whole function which can be re-borrowed from the public pos.func member.

Fields
func: &'f mut Function

The referenced function.

isa: &'f dyn TargetIsa

The target ISA that will be used to encode instructions.

Implementations

impl<'f> EncCursor<'f>

pub fn new(func: &'f mut Function, isa: &'f dyn TargetIsa) -> Self

Create a new EncCursor pointing nowhere.

pub fn use_srcloc(&mut self, inst: Inst)

Use the source location of inst for future instructions.

pub fn ins(&mut self) -> InsertBuilder<'_, &mut EncCursor<'f>>

Create an instruction builder that will insert an encoded instruction at the current position.

The builder will panic if it is used to insert an instruction that can’t be encoded for self.isa.

pub fn built_inst(&self) -> Inst

Get the last built instruction.

This returns the last instruction that was built using the ins() method on this cursor. Panics if no instruction was built.

pub fn display_inst(&self, inst: Inst) -> DisplayInst<'_>

Return an object that can display inst.

This is a convenience wrapper for the DFG equivalent.