cranelift_codegen

Enum CodegenError

pub enum CodegenError {
    Verifier(VerifierErrors),
    ImplLimitExceeded,
    CodeTooLarge,
    Unsupported(String),
    RegisterMappingError(RegisterMappingError),
    Regalloc(CheckerErrors),
    Pcc(PccError),
}

A compilation error.

When Cranelift fails to compile a function, it will return one of these error codes.

Variants

Verifier(VerifierErrors)

A list of IR verifier errors.

This always represents a bug, either in the code that generated IR for Cranelift, or a bug in Cranelift itself.

ImplLimitExceeded

An implementation limit was exceeded.

Cranelift can compile very large and complicated functions, but the implementation has limits that cause compilation to fail when they are exceeded.

CodeTooLarge

The code size for the function is too large.

Different target ISAs may impose a limit on the size of a compiled function. If that limit is exceeded, compilation fails.

Unsupported(String)

Something is not supported by the code generator. This might be an indication that a feature is used without explicitly enabling it, or that something is temporarily unsupported by a given target backend.

RegisterMappingError(RegisterMappingError)

A failure to map Cranelift register representation to a DWARF register representation.

Regalloc(CheckerErrors)

Register allocator internal error discovered by the symbolic checker.

Pcc(PccError)

Proof-carrying-code validation error.