cranelift_codegen::ir::function

Struct Function

pub struct Function {
    pub name: UserFuncName,
    pub stencil: FunctionStencil,
    pub params: FunctionParameters,
}

Functions can be cloned, but it is not a very fast operation. The clone will have all the same entity numbers as the original.

Fields
name: UserFuncName

Name of this function.

Mostly used by .clif files, only there for debugging / naming purposes.

stencil: FunctionStencil

All the fields required for compiling a function, independently of details irrelevant to compilation and that are stored in the FunctionParameters params field instead.

params: FunctionParameters

All the parameters that can be applied onto the function stencil, that is, that don’t matter when caching compilation artifacts.

Implementations

impl Function

pub fn with_name_signature(name: UserFuncName, sig: Signature) -> Self

Create a function with the given name and signature.

pub fn clear(&mut self)

Clear all data structures in this function.

pub fn new() -> Self

Create a new empty, anonymous function with a Fast calling convention.

pub fn display(&self) -> DisplayFunction<'_>

Return an object that can display this function with correct ISA-specific annotations.

pub fn set_srcloc(&mut self, inst: Inst, srcloc: SourceLoc)

Sets an absolute source location for the given instruction.

If no base source location has been set yet, records it at the same time.

pub fn srcloc(&self, inst: Inst) -> SourceLoc

Returns an absolute source location for the given instruction.

pub fn declare_imported_user_function( &mut self, name: UserExternalName, ) -> UserExternalNameRef

Declare a user-defined external function import, to be referenced in ExtFuncData::User later.

pub fn import_function(&mut self, data: ExtFuncData) -> FuncRef

Declare an external function import.