Module cranelift_codegen::ir::immediates

Immediate operands for Cranelift instructions

This module defines the types of immediate operands that can appear on Cranelift instructions. Each type here should have a corresponding definition in the cranelift-codegen/meta/src/shared/immediates crate in the meta language.

Structs

Ieee32

An IEEE binary32 immediate floating point value, represented as a u32 containing the bit pattern.

Ieee64

An IEEE binary64 immediate floating point value, represented as a u64 containing the bit pattern.

Imm64

64-bit immediate signed integer operand.

Offset32

32-bit signed immediate offset.

Uimm32

A 32-bit unsigned integer immediate operand.

Uimm64

64-bit immediate unsigned integer operand.

V128Imm

A 128-bit immediate operand.

Traits

IntoBytes

Convert a type into a vector of bytes; all implementors in this file must use little-endian orderings of bytes to match WebAssembly’s little-endianness.

Type Definitions

Uimm8

8-bit unsigned integer immediate operand.