Enum cranelift_codegen::ir::instructions::Opcode

#[repr(u8)]
pub enum Opcode {

    Jump = 1,
    Brif = 2,
    BrTable = 3,
    Debugtrap = 4,
    Trap = 5,
    Trapz = 6,
    ResumableTrap = 7,
    Trapnz = 8,
    ResumableTrapnz = 9,
    Return = 10,
    Call = 11,
    CallIndirect = 12,
    ReturnCall = 13,
    ReturnCallIndirect = 14,
    FuncAddr = 15,
    Splat = 16,
    Swizzle = 17,
    X86Pshufb = 18,
    Insertlane = 19,
    Extractlane = 20,
    Smin = 21,
    Umin = 22,
    Smax = 23,
    Umax = 24,
    AvgRound = 25,
    UaddSat = 26,
    SaddSat = 27,
    UsubSat = 28,
    SsubSat = 29,
    Load = 30,
    Store = 31,
    Uload8 = 32,
    Sload8 = 33,
    Istore8 = 34,
    Uload16 = 35,
    Sload16 = 36,
    Istore16 = 37,
    Uload32 = 38,
    Sload32 = 39,
    Istore32 = 40,
    Uload8x8 = 41,
    Sload8x8 = 42,
    Uload16x4 = 43,
    Sload16x4 = 44,
    Uload32x2 = 45,
    Sload32x2 = 46,
    StackLoad = 47,
    StackStore = 48,
    StackAddr = 49,
    DynamicStackLoad = 50,
    DynamicStackStore = 51,
    DynamicStackAddr = 52,
    GlobalValue = 53,
    SymbolValue = 54,
    TlsValue = 55,
    GetPinnedReg = 56,
    SetPinnedReg = 57,
    GetFramePointer = 58,
    GetStackPointer = 59,
    GetReturnAddress = 60,
    Iconst = 61,
    F32const = 62,
    F64const = 63,
    Vconst = 64,
    Shuffle = 65,
    Null = 66,
    Nop = 67,
    Select = 68,
    SelectSpectreGuard = 69,
    Bitselect = 70,
    X86Blendv = 71,
    VanyTrue = 72,
    VallTrue = 73,
    VhighBits = 74,
    Icmp = 75,
    IcmpImm = 76,
    Iadd = 77,
    Isub = 78,
    Ineg = 79,
    Iabs = 80,
    Imul = 81,
    Umulhi = 82,
    Smulhi = 83,
    SqmulRoundSat = 84,
    X86Pmulhrsw = 85,
    Udiv = 86,
    Sdiv = 87,
    Urem = 88,
    Srem = 89,
    IaddImm = 90,
    ImulImm = 91,
    UdivImm = 92,
    SdivImm = 93,
    UremImm = 94,
    SremImm = 95,
    IrsubImm = 96,
    IaddCin = 97,
    IaddCarry = 98,
    UaddOverflow = 99,
    SaddOverflow = 100,
    UsubOverflow = 101,
    SsubOverflow = 102,
    UmulOverflow = 103,
    SmulOverflow = 104,
    UaddOverflowTrap = 105,
    IsubBin = 106,
    IsubBorrow = 107,
    Band = 108,
    Bor = 109,
    Bxor = 110,
    Bnot = 111,
    BandNot = 112,
    BorNot = 113,
    BxorNot = 114,
    BandImm = 115,
    BorImm = 116,
    BxorImm = 117,
    Rotl = 118,
    Rotr = 119,
    RotlImm = 120,
    RotrImm = 121,
    Ishl = 122,
    Ushr = 123,
    Sshr = 124,
    IshlImm = 125,
    UshrImm = 126,
    SshrImm = 127,
    Bitrev = 128,
    Clz = 129,
    Cls = 130,
    Ctz = 131,
    Bswap = 132,
    Popcnt = 133,
    Fcmp = 134,
    Fadd = 135,
    Fsub = 136,
    Fmul = 137,
    Fdiv = 138,
    Sqrt = 139,
    Fma = 140,
    Fneg = 141,
    Fabs = 142,
    Fcopysign = 143,
    Fmin = 144,
    Fmax = 145,
    Ceil = 146,
    Floor = 147,
    Trunc = 148,
    Nearest = 149,
    IsNull = 150,
    IsInvalid = 151,
    Bitcast = 152,
    ScalarToVector = 153,
    Bmask = 154,
    Ireduce = 155,
    Snarrow = 156,
    Unarrow = 157,
    Uunarrow = 158,
    SwidenLow = 159,
    SwidenHigh = 160,
    UwidenLow = 161,
    UwidenHigh = 162,
    IaddPairwise = 163,
    X86Pmaddubsw = 164,
    Uextend = 165,
    Sextend = 166,
    Fpromote = 167,
    Fdemote = 168,
    Fvdemote = 169,
    FvpromoteLow = 170,
    FcvtToUint = 171,
    FcvtToSint = 172,
    FcvtToUintSat = 173,
    FcvtToSintSat = 174,
    X86Cvtt2dq = 175,
    FcvtFromUint = 176,
    FcvtFromSint = 177,
    Isplit = 178,
    Iconcat = 179,
    AtomicRmw = 180,
    AtomicCas = 181,
    AtomicLoad = 182,
    AtomicStore = 183,
    Fence = 184,
    ExtractVector = 185,

}

An instruction opcode.

All instructions from all supported ISAs are present.

Variants

Jump = 1

jump block_call. (Jump)

Brif = 2

brif c, block_then, block_else. (Brif) Type inferred from c.

BrTable = 3

br_table x, JT. (BranchTable)

Debugtrap = 4

debugtrap. (NullAry)

Trap = 5

trap code. (Trap)

Trapz = 6

trapz c, code. (CondTrap) Type inferred from c.

ResumableTrap = 7

resumable_trap code. (Trap)

Trapnz = 8

trapnz c, code. (CondTrap) Type inferred from c.

ResumableTrapnz = 9

resumable_trapnz c, code. (CondTrap) Type inferred from c.

Return = 10

return rvals. (MultiAry)

Call = 11

rvals = call FN, args. (Call)

CallIndirect = 12

rvals = call_indirect SIG, callee, args. (CallIndirect) Type inferred from callee.

ReturnCall = 13

return_call FN, args. (Call)

ReturnCallIndirect = 14

return_call_indirect SIG, callee, args. (CallIndirect) Type inferred from callee.

FuncAddr = 15

addr = func_addr FN. (FuncAddr)

Splat = 16

a = splat x. (Unary)

Swizzle = 17

a = swizzle x, y. (Binary)

X86Pshufb = 18

a = x86_pshufb x, y. (Binary)

Insertlane = 19

a = insertlane x, y, Idx. (TernaryImm8) Type inferred from x.

Extractlane = 20

a = extractlane x, Idx. (BinaryImm8) Type inferred from x.

Smin = 21

a = smin x, y. (Binary) Type inferred from x.

Umin = 22

a = umin x, y. (Binary) Type inferred from x.

Smax = 23

a = smax x, y. (Binary) Type inferred from x.

Umax = 24

a = umax x, y. (Binary) Type inferred from x.

AvgRound = 25

a = avg_round x, y. (Binary) Type inferred from x.

UaddSat = 26

a = uadd_sat x, y. (Binary) Type inferred from x.

SaddSat = 27

a = sadd_sat x, y. (Binary) Type inferred from x.

UsubSat = 28

a = usub_sat x, y. (Binary) Type inferred from x.

SsubSat = 29

a = ssub_sat x, y. (Binary) Type inferred from x.

Load = 30

a = load MemFlags, p, Offset. (Load)

Store = 31

store MemFlags, x, p, Offset. (Store) Type inferred from x.

Uload8 = 32

a = uload8 MemFlags, p, Offset. (Load)

Sload8 = 33

a = sload8 MemFlags, p, Offset. (Load)

Istore8 = 34

istore8 MemFlags, x, p, Offset. (Store) Type inferred from x.

Uload16 = 35

a = uload16 MemFlags, p, Offset. (Load)

Sload16 = 36

a = sload16 MemFlags, p, Offset. (Load)

Istore16 = 37

istore16 MemFlags, x, p, Offset. (Store) Type inferred from x.

Uload32 = 38

a = uload32 MemFlags, p, Offset. (Load) Type inferred from p.

Sload32 = 39

a = sload32 MemFlags, p, Offset. (Load) Type inferred from p.

Istore32 = 40

istore32 MemFlags, x, p, Offset. (Store) Type inferred from x.

Uload8x8 = 41

a = uload8x8 MemFlags, p, Offset. (Load) Type inferred from p.

Sload8x8 = 42

a = sload8x8 MemFlags, p, Offset. (Load) Type inferred from p.

Uload16x4 = 43

a = uload16x4 MemFlags, p, Offset. (Load) Type inferred from p.

Sload16x4 = 44

a = sload16x4 MemFlags, p, Offset. (Load) Type inferred from p.

Uload32x2 = 45

a = uload32x2 MemFlags, p, Offset. (Load) Type inferred from p.

Sload32x2 = 46

a = sload32x2 MemFlags, p, Offset. (Load) Type inferred from p.

StackLoad = 47

a = stack_load SS, Offset. (StackLoad)

StackStore = 48

stack_store x, SS, Offset. (StackStore) Type inferred from x.

StackAddr = 49

addr = stack_addr SS, Offset. (StackLoad)

DynamicStackLoad = 50

a = dynamic_stack_load DSS. (DynamicStackLoad)

DynamicStackStore = 51

dynamic_stack_store x, DSS. (DynamicStackStore) Type inferred from x.

DynamicStackAddr = 52

addr = dynamic_stack_addr DSS. (DynamicStackLoad)

GlobalValue = 53

a = global_value GV. (UnaryGlobalValue)

SymbolValue = 54

a = symbol_value GV. (UnaryGlobalValue)

TlsValue = 55

a = tls_value GV. (UnaryGlobalValue)

GetPinnedReg = 56

addr = get_pinned_reg. (NullAry)

SetPinnedReg = 57

set_pinned_reg addr. (Unary) Type inferred from addr.

GetFramePointer = 58

addr = get_frame_pointer. (NullAry)

GetStackPointer = 59

addr = get_stack_pointer. (NullAry)

GetReturnAddress = 60

addr = get_return_address. (NullAry)

Iconst = 61

a = iconst N. (UnaryImm)

F32const = 62

a = f32const N. (UnaryIeee32)

F64const = 63

a = f64const N. (UnaryIeee64)

Vconst = 64

a = vconst N. (UnaryConst)

Shuffle = 65

a = shuffle a, b, mask. (Shuffle)

Null = 66

a = null. (NullAry)

Nop = 67

nop. (NullAry)

Select = 68

a = select c, x, y. (Ternary) Type inferred from x.

SelectSpectreGuard = 69

a = select_spectre_guard c, x, y. (Ternary) Type inferred from x.

Bitselect = 70

a = bitselect c, x, y. (Ternary) Type inferred from x.

X86Blendv = 71

a = x86_blendv c, x, y. (Ternary) Type inferred from x.

VanyTrue = 72

s = vany_true a. (Unary) Type inferred from a.

VallTrue = 73

s = vall_true a. (Unary) Type inferred from a.

VhighBits = 74

x = vhigh_bits a. (Unary)

Icmp = 75

a = icmp Cond, x, y. (IntCompare) Type inferred from x.

IcmpImm = 76

a = icmp_imm Cond, x, Y. (IntCompareImm) Type inferred from x.

Iadd = 77

a = iadd x, y. (Binary) Type inferred from x.

Isub = 78

a = isub x, y. (Binary) Type inferred from x.

Ineg = 79

a = ineg x. (Unary) Type inferred from x.

Iabs = 80

a = iabs x. (Unary) Type inferred from x.

Imul = 81

a = imul x, y. (Binary) Type inferred from x.

Umulhi = 82

a = umulhi x, y. (Binary) Type inferred from x.

Smulhi = 83

a = smulhi x, y. (Binary) Type inferred from x.

SqmulRoundSat = 84

a = sqmul_round_sat x, y. (Binary) Type inferred from x.

X86Pmulhrsw = 85

a = x86_pmulhrsw x, y. (Binary) Type inferred from x.

Udiv = 86

a = udiv x, y. (Binary) Type inferred from x.

Sdiv = 87

a = sdiv x, y. (Binary) Type inferred from x.

Urem = 88

a = urem x, y. (Binary) Type inferred from x.

Srem = 89

a = srem x, y. (Binary) Type inferred from x.

IaddImm = 90

a = iadd_imm x, Y. (BinaryImm64) Type inferred from x.

ImulImm = 91

a = imul_imm x, Y. (BinaryImm64) Type inferred from x.

UdivImm = 92

a = udiv_imm x, Y. (BinaryImm64) Type inferred from x.

SdivImm = 93

a = sdiv_imm x, Y. (BinaryImm64) Type inferred from x.

UremImm = 94

a = urem_imm x, Y. (BinaryImm64) Type inferred from x.

SremImm = 95

a = srem_imm x, Y. (BinaryImm64) Type inferred from x.

IrsubImm = 96

a = irsub_imm x, Y. (BinaryImm64) Type inferred from x.

IaddCin = 97

a = iadd_cin x, y, c_in. (Ternary) Type inferred from y.

IaddCarry = 98

a, c_out = iadd_carry x, y, c_in. (Ternary) Type inferred from y.

UaddOverflow = 99

a, of = uadd_overflow x, y. (Binary) Type inferred from x.

SaddOverflow = 100

a, of = sadd_overflow x, y. (Binary) Type inferred from x.

UsubOverflow = 101

a, of = usub_overflow x, y. (Binary) Type inferred from x.

SsubOverflow = 102

a, of = ssub_overflow x, y. (Binary) Type inferred from x.

UmulOverflow = 103

a, of = umul_overflow x, y. (Binary) Type inferred from x.

SmulOverflow = 104

a, of = smul_overflow x, y. (Binary) Type inferred from x.

UaddOverflowTrap = 105

a = uadd_overflow_trap x, y, code. (IntAddTrap) Type inferred from x.

IsubBin = 106

a = isub_bin x, y, b_in. (Ternary) Type inferred from y.

IsubBorrow = 107

a, b_out = isub_borrow x, y, b_in. (Ternary) Type inferred from y.

Band = 108

a = band x, y. (Binary) Type inferred from x.

Bor = 109

a = bor x, y. (Binary) Type inferred from x.

Bxor = 110

a = bxor x, y. (Binary) Type inferred from x.

Bnot = 111

a = bnot x. (Unary) Type inferred from x.

BandNot = 112

a = band_not x, y. (Binary) Type inferred from x.

BorNot = 113

a = bor_not x, y. (Binary) Type inferred from x.

BxorNot = 114

a = bxor_not x, y. (Binary) Type inferred from x.

BandImm = 115

a = band_imm x, Y. (BinaryImm64) Type inferred from x.

BorImm = 116

a = bor_imm x, Y. (BinaryImm64) Type inferred from x.

BxorImm = 117

a = bxor_imm x, Y. (BinaryImm64) Type inferred from x.

Rotl = 118

a = rotl x, y. (Binary) Type inferred from x.

Rotr = 119

a = rotr x, y. (Binary) Type inferred from x.

RotlImm = 120

a = rotl_imm x, Y. (BinaryImm64) Type inferred from x.

RotrImm = 121

a = rotr_imm x, Y. (BinaryImm64) Type inferred from x.

Ishl = 122

a = ishl x, y. (Binary) Type inferred from x.

Ushr = 123

a = ushr x, y. (Binary) Type inferred from x.

Sshr = 124

a = sshr x, y. (Binary) Type inferred from x.

IshlImm = 125

a = ishl_imm x, Y. (BinaryImm64) Type inferred from x.

UshrImm = 126

a = ushr_imm x, Y. (BinaryImm64) Type inferred from x.

SshrImm = 127

a = sshr_imm x, Y. (BinaryImm64) Type inferred from x.

Bitrev = 128

a = bitrev x. (Unary) Type inferred from x.

Clz = 129

a = clz x. (Unary) Type inferred from x.

Cls = 130

a = cls x. (Unary) Type inferred from x.

Ctz = 131

a = ctz x. (Unary) Type inferred from x.

Bswap = 132

a = bswap x. (Unary) Type inferred from x.

Popcnt = 133

a = popcnt x. (Unary) Type inferred from x.

Fcmp = 134

a = fcmp Cond, x, y. (FloatCompare) Type inferred from x.

Fadd = 135

a = fadd x, y. (Binary) Type inferred from x.

Fsub = 136

a = fsub x, y. (Binary) Type inferred from x.

Fmul = 137

a = fmul x, y. (Binary) Type inferred from x.

Fdiv = 138

a = fdiv x, y. (Binary) Type inferred from x.

Sqrt = 139

a = sqrt x. (Unary) Type inferred from x.

Fma = 140

a = fma x, y, z. (Ternary) Type inferred from y.

Fneg = 141

a = fneg x. (Unary) Type inferred from x.

Fabs = 142

a = fabs x. (Unary) Type inferred from x.

Fcopysign = 143

a = fcopysign x, y. (Binary) Type inferred from x.

Fmin = 144

a = fmin x, y. (Binary) Type inferred from x.

Fmax = 145

a = fmax x, y. (Binary) Type inferred from x.

Ceil = 146

a = ceil x. (Unary) Type inferred from x.

Floor = 147

a = floor x. (Unary) Type inferred from x.

Trunc = 148

a = trunc x. (Unary) Type inferred from x.

Nearest = 149

a = nearest x. (Unary) Type inferred from x.

IsNull = 150

a = is_null x. (Unary) Type inferred from x.

IsInvalid = 151

a = is_invalid x. (Unary) Type inferred from x.

Bitcast = 152

a = bitcast MemFlags, x. (LoadNoOffset)

ScalarToVector = 153

a = scalar_to_vector s. (Unary)

Bmask = 154

a = bmask x. (Unary)

Ireduce = 155

a = ireduce x. (Unary)

Snarrow = 156

a = snarrow x, y. (Binary) Type inferred from x.

Unarrow = 157

a = unarrow x, y. (Binary) Type inferred from x.

Uunarrow = 158

a = uunarrow x, y. (Binary) Type inferred from x.

SwidenLow = 159

a = swiden_low x. (Unary) Type inferred from x.

SwidenHigh = 160

a = swiden_high x. (Unary) Type inferred from x.

UwidenLow = 161

a = uwiden_low x. (Unary) Type inferred from x.

UwidenHigh = 162

a = uwiden_high x. (Unary) Type inferred from x.

IaddPairwise = 163

a = iadd_pairwise x, y. (Binary) Type inferred from x.

X86Pmaddubsw = 164

a = x86_pmaddubsw x, y. (Binary)

Uextend = 165

a = uextend x. (Unary)

Sextend = 166

a = sextend x. (Unary)

Fpromote = 167

a = fpromote x. (Unary)

Fdemote = 168

a = fdemote x. (Unary)

Fvdemote = 169

a = fvdemote x. (Unary)

FvpromoteLow = 170

x = fvpromote_low a. (Unary)

FcvtToUint = 171

a = fcvt_to_uint x. (Unary)

FcvtToSint = 172

a = fcvt_to_sint x. (Unary)

FcvtToUintSat = 173

a = fcvt_to_uint_sat x. (Unary)

FcvtToSintSat = 174

a = fcvt_to_sint_sat x. (Unary)

X86Cvtt2dq = 175

a = x86_cvtt2dq x. (Unary)

FcvtFromUint = 176

a = fcvt_from_uint x. (Unary)

FcvtFromSint = 177

a = fcvt_from_sint x. (Unary)

Isplit = 178

lo, hi = isplit x. (Unary) Type inferred from x.

Iconcat = 179

a = iconcat lo, hi. (Binary) Type inferred from lo.

AtomicRmw = 180

a = atomic_rmw MemFlags, AtomicRmwOp, p, x. (AtomicRmw)

AtomicCas = 181

a = atomic_cas MemFlags, p, e, x. (AtomicCas) Type inferred from x.

AtomicLoad = 182

a = atomic_load MemFlags, p. (LoadNoOffset)

AtomicStore = 183

atomic_store MemFlags, x, p. (StoreNoOffset) Type inferred from x.

Fence = 184

fence. (NullAry)

ExtractVector = 185

a = extract_vector x, y. (BinaryImm8) Type inferred from x.

Implementations

impl Opcode

pub fn is_terminator(self) -> bool

True for instructions that terminate the block

pub fn is_branch(self) -> bool

True for all branch or jump instructions.

pub fn is_call(self) -> bool

Is this a call instruction?

pub fn is_return(self) -> bool

Is this a return instruction?

pub fn can_load(self) -> bool

Can this instruction read from memory?

pub fn can_store(self) -> bool

Can this instruction write to memory?

pub fn can_trap(self) -> bool

Can this instruction cause a trap?

pub fn other_side_effects(self) -> bool

Does this instruction have other side effects besides can_* flags?

pub fn side_effects_idempotent(self) -> bool

Despite having side effects, is this instruction okay to GVN?

pub fn all() -> &'static [Opcode]

All cranelift opcodes.

impl Opcode

pub fn format(self) -> InstructionFormat

Get the instruction format for this opcode.

pub fn constraints(self) -> OpcodeConstraints

Get the constraint descriptor for this opcode. Panic if this is called on NotAnOpcode.

pub fn is_resumable_trap(&self) -> bool

Returns true if the instruction is a resumable trap.