Enum cranelift_codegen::ir::instructions::Opcode

#[repr(u16)]
pub enum Opcode {

    Jump = 1,
    Brz = 2,
    Brnz = 3,
    BrIcmp = 4,
    Brif = 5,
    Brff = 6,
    BrTable = 7,
    Debugtrap = 8,
    Trap = 9,
    Trapz = 10,
    ResumableTrap = 11,
    Trapnz = 12,
    ResumableTrapnz = 13,
    Trapif = 14,
    Trapff = 15,
    Return = 16,
    Call = 17,
    CallIndirect = 18,
    FuncAddr = 19,
    Splat = 20,
    Swizzle = 21,
    Insertlane = 22,
    Extractlane = 23,
    Imin = 24,
    Umin = 25,
    Imax = 26,
    Umax = 27,
    AvgRound = 28,
    UaddSat = 29,
    SaddSat = 30,
    UsubSat = 31,
    SsubSat = 32,
    Load = 33,
    Store = 34,
    Uload8 = 35,
    Sload8 = 36,
    Istore8 = 37,
    Uload16 = 38,
    Sload16 = 39,
    Istore16 = 40,
    Uload32 = 41,
    Sload32 = 42,
    Istore32 = 43,
    Uload8x8 = 44,
    Sload8x8 = 45,
    Uload16x4 = 46,
    Sload16x4 = 47,
    Uload32x2 = 48,
    Sload32x2 = 49,
    StackLoad = 50,
    StackStore = 51,
    StackAddr = 52,
    DynamicStackLoad = 53,
    DynamicStackStore = 54,
    DynamicStackAddr = 55,
    GlobalValue = 56,
    SymbolValue = 57,
    TlsValue = 58,
    HeapAddr = 59,
    GetPinnedReg = 60,
    SetPinnedReg = 61,
    GetFramePointer = 62,
    GetStackPointer = 63,
    GetReturnAddress = 64,
    TableAddr = 65,
    Iconst = 66,
    F32const = 67,
    F64const = 68,
    Bconst = 69,
    Vconst = 70,
    Shuffle = 71,
    Null = 72,
    Nop = 73,
    Select = 74,
    Selectif = 75,
    SelectifSpectreGuard = 76,
    Bitselect = 77,
    Copy = 78,
    Vsplit = 79,
    Vconcat = 80,
    Vselect = 81,
    VanyTrue = 82,
    VallTrue = 83,
    VhighBits = 84,
    Icmp = 85,
    IcmpImm = 86,
    Ifcmp = 87,
    IfcmpImm = 88,
    Iadd = 89,
    Isub = 90,
    Ineg = 91,
    Iabs = 92,
    Imul = 93,
    Umulhi = 94,
    Smulhi = 95,
    SqmulRoundSat = 96,
    Udiv = 97,
    Sdiv = 98,
    Urem = 99,
    Srem = 100,
    IaddImm = 101,
    ImulImm = 102,
    UdivImm = 103,
    SdivImm = 104,
    UremImm = 105,
    SremImm = 106,
    IrsubImm = 107,
    IaddCin = 108,
    IaddIfcin = 109,
    IaddCout = 110,
    IaddIfcout = 111,
    IaddCarry = 112,
    IaddIfcarry = 113,
    IsubBin = 114,
    IsubIfbin = 115,
    IsubBout = 116,
    IsubIfbout = 117,
    IsubBorrow = 118,
    IsubIfborrow = 119,
    Band = 120,
    Bor = 121,
    Bxor = 122,
    Bnot = 123,
    BandNot = 124,
    BorNot = 125,
    BxorNot = 126,
    BandImm = 127,
    BorImm = 128,
    BxorImm = 129,
    Rotl = 130,
    Rotr = 131,
    RotlImm = 132,
    RotrImm = 133,
    Ishl = 134,
    Ushr = 135,
    Sshr = 136,
    IshlImm = 137,
    UshrImm = 138,
    SshrImm = 139,
    Bitrev = 140,
    Clz = 141,
    Cls = 142,
    Ctz = 143,
    Popcnt = 144,
    Fcmp = 145,
    Ffcmp = 146,
    Fadd = 147,
    Fsub = 148,
    Fmul = 149,
    Fdiv = 150,
    Sqrt = 151,
    Fma = 152,
    Fneg = 153,
    Fabs = 154,
    Fcopysign = 155,
    Fmin = 156,
    FminPseudo = 157,
    Fmax = 158,
    FmaxPseudo = 159,
    Ceil = 160,
    Floor = 161,
    Trunc = 162,
    Nearest = 163,
    IsNull = 164,
    IsInvalid = 165,
    Trueif = 166,
    Trueff = 167,
    Bitcast = 168,
    RawBitcast = 169,
    ScalarToVector = 170,
    Breduce = 171,
    Bextend = 172,
    Bint = 173,
    Bmask = 174,
    Ireduce = 175,
    Snarrow = 176,
    Unarrow = 177,
    Uunarrow = 178,
    SwidenLow = 179,
    SwidenHigh = 180,
    UwidenLow = 181,
    UwidenHigh = 182,
    IaddPairwise = 183,
    WideningPairwiseDotProductS = 184,
    Uextend = 185,
    Sextend = 186,
    Fpromote = 187,
    Fdemote = 188,
    Fvdemote = 189,
    FvpromoteLow = 190,
    FcvtToUint = 191,
    FcvtToSint = 192,
    FcvtToUintSat = 193,
    FcvtToSintSat = 194,
    FcvtFromUint = 195,
    FcvtFromSint = 196,
    FcvtLowFromSint = 197,
    Isplit = 198,
    Iconcat = 199,
    AtomicRmw = 200,
    AtomicCas = 201,
    AtomicLoad = 202,
    AtomicStore = 203,
    Fence = 204,
    ExtractVector = 205,

}

An instruction opcode.

All instructions from all supported ISAs are present.

Variants

Jump = 1

jump block, args. (Jump)

Brz = 2

brz c, block, args. (Branch) Type inferred from c.

Brnz = 3

brnz c, block, args. (Branch) Type inferred from c.

BrIcmp = 4

br_icmp Cond, x, y, block, args. (BranchIcmp) Type inferred from x.

Brif = 5

brif Cond, f, block, args. (BranchInt)

Brff = 6

brff Cond, f, block, args. (BranchFloat)

BrTable = 7

br_table x, block, JT. (BranchTable)

Debugtrap = 8

debugtrap. (NullAry)

Trap = 9

trap code. (Trap)

Trapz = 10

trapz c, code. (CondTrap) Type inferred from c.

ResumableTrap = 11

resumable_trap code. (Trap)

Trapnz = 12

trapnz c, code. (CondTrap) Type inferred from c.

ResumableTrapnz = 13

resumable_trapnz c, code. (CondTrap) Type inferred from c.

Trapif = 14

trapif Cond, f, code. (IntCondTrap)

Trapff = 15

trapff Cond, f, code. (FloatCondTrap)

Return = 16

return rvals. (MultiAry)

Call = 17

rvals = call FN, args. (Call)

CallIndirect = 18

rvals = call_indirect SIG, callee, args. (CallIndirect) Type inferred from callee.

FuncAddr = 19

addr = func_addr FN. (FuncAddr)

Splat = 20

a = splat x. (Unary)

Swizzle = 21

a = swizzle x, y. (Binary)

Insertlane = 22

a = insertlane x, y, Idx. (TernaryImm8) Type inferred from x.

Extractlane = 23

a = extractlane x, Idx. (BinaryImm8) Type inferred from x.

Imin = 24

a = imin x, y. (Binary) Type inferred from x.

Umin = 25

a = umin x, y. (Binary) Type inferred from x.

Imax = 26

a = imax x, y. (Binary) Type inferred from x.

Umax = 27

a = umax x, y. (Binary) Type inferred from x.

AvgRound = 28

a = avg_round x, y. (Binary) Type inferred from x.

UaddSat = 29

a = uadd_sat x, y. (Binary) Type inferred from x.

SaddSat = 30

a = sadd_sat x, y. (Binary) Type inferred from x.

UsubSat = 31

a = usub_sat x, y. (Binary) Type inferred from x.

SsubSat = 32

a = ssub_sat x, y. (Binary) Type inferred from x.

Load = 33

a = load MemFlags, p, Offset. (Load)

Store = 34

store MemFlags, x, p, Offset. (Store) Type inferred from x.

Uload8 = 35

a = uload8 MemFlags, p, Offset. (Load)

Sload8 = 36

a = sload8 MemFlags, p, Offset. (Load)

Istore8 = 37

istore8 MemFlags, x, p, Offset. (Store) Type inferred from x.

Uload16 = 38

a = uload16 MemFlags, p, Offset. (Load)

Sload16 = 39

a = sload16 MemFlags, p, Offset. (Load)

Istore16 = 40

istore16 MemFlags, x, p, Offset. (Store) Type inferred from x.

Uload32 = 41

a = uload32 MemFlags, p, Offset. (Load) Type inferred from p.

Sload32 = 42

a = sload32 MemFlags, p, Offset. (Load) Type inferred from p.

Istore32 = 43

istore32 MemFlags, x, p, Offset. (Store) Type inferred from x.

Uload8x8 = 44

a = uload8x8 MemFlags, p, Offset. (Load) Type inferred from p.

Sload8x8 = 45

a = sload8x8 MemFlags, p, Offset. (Load) Type inferred from p.

Uload16x4 = 46

a = uload16x4 MemFlags, p, Offset. (Load) Type inferred from p.

Sload16x4 = 47

a = sload16x4 MemFlags, p, Offset. (Load) Type inferred from p.

Uload32x2 = 48

a = uload32x2 MemFlags, p, Offset. (Load) Type inferred from p.

Sload32x2 = 49

a = sload32x2 MemFlags, p, Offset. (Load) Type inferred from p.

StackLoad = 50

a = stack_load SS, Offset. (StackLoad)

StackStore = 51

stack_store x, SS, Offset. (StackStore) Type inferred from x.

StackAddr = 52

addr = stack_addr SS, Offset. (StackLoad)

DynamicStackLoad = 53

a = dynamic_stack_load DSS. (DynamicStackLoad)

DynamicStackStore = 54

dynamic_stack_store x, DSS. (DynamicStackStore) Type inferred from x.

DynamicStackAddr = 55

addr = dynamic_stack_addr DSS. (DynamicStackLoad)

GlobalValue = 56

a = global_value GV. (UnaryGlobalValue)

SymbolValue = 57

a = symbol_value GV. (UnaryGlobalValue)

TlsValue = 58

a = tls_value GV. (UnaryGlobalValue)

HeapAddr = 59

addr = heap_addr H, p, Size. (HeapAddr)

GetPinnedReg = 60

addr = get_pinned_reg. (NullAry)

SetPinnedReg = 61

set_pinned_reg addr. (Unary) Type inferred from addr.

GetFramePointer = 62

addr = get_frame_pointer. (NullAry)

GetStackPointer = 63

addr = get_stack_pointer. (NullAry)

GetReturnAddress = 64

addr = get_return_address. (NullAry)

TableAddr = 65

addr = table_addr T, p, Offset. (TableAddr)

Iconst = 66

a = iconst N. (UnaryImm)

F32const = 67

a = f32const N. (UnaryIeee32)

F64const = 68

a = f64const N. (UnaryIeee64)

Bconst = 69

a = bconst N. (UnaryBool)

Vconst = 70

a = vconst N. (UnaryConst)

Shuffle = 71

a = shuffle a, b, mask. (Shuffle) Type inferred from a.

Null = 72

a = null. (NullAry)

Nop = 73

nop. (NullAry)

Select = 74

a = select c, x, y. (Ternary) Type inferred from x.

Selectif = 75

a = selectif cc, flags, x, y. (IntSelect)

SelectifSpectreGuard = 76

a = selectif_spectre_guard cc, flags, x, y. (IntSelect)

Bitselect = 77

a = bitselect c, x, y. (Ternary) Type inferred from x.

Copy = 78

a = copy x. (Unary) Type inferred from x.

Vsplit = 79

lo, hi = vsplit x. (Unary) Type inferred from x.

Vconcat = 80

a = vconcat x, y. (Binary) Type inferred from x.

Vselect = 81

a = vselect c, x, y. (Ternary) Type inferred from x.

VanyTrue = 82

s = vany_true a. (Unary) Type inferred from a.

VallTrue = 83

s = vall_true a. (Unary) Type inferred from a.

VhighBits = 84

x = vhigh_bits a. (Unary)

Icmp = 85

a = icmp Cond, x, y. (IntCompare) Type inferred from x.

IcmpImm = 86

a = icmp_imm Cond, x, Y. (IntCompareImm) Type inferred from x.

Ifcmp = 87

f = ifcmp x, y. (Binary) Type inferred from x.

IfcmpImm = 88

f = ifcmp_imm x, Y. (BinaryImm64) Type inferred from x.

Iadd = 89

a = iadd x, y. (Binary) Type inferred from x.

Isub = 90

a = isub x, y. (Binary) Type inferred from x.

Ineg = 91

a = ineg x. (Unary) Type inferred from x.

Iabs = 92

a = iabs x. (Unary) Type inferred from x.

Imul = 93

a = imul x, y. (Binary) Type inferred from x.

Umulhi = 94

a = umulhi x, y. (Binary) Type inferred from x.

Smulhi = 95

a = smulhi x, y. (Binary) Type inferred from x.

SqmulRoundSat = 96

a = sqmul_round_sat x, y. (Binary) Type inferred from x.

Udiv = 97

a = udiv x, y. (Binary) Type inferred from x.

Sdiv = 98

a = sdiv x, y. (Binary) Type inferred from x.

Urem = 99

a = urem x, y. (Binary) Type inferred from x.

Srem = 100

a = srem x, y. (Binary) Type inferred from x.

IaddImm = 101

a = iadd_imm x, Y. (BinaryImm64) Type inferred from x.

ImulImm = 102

a = imul_imm x, Y. (BinaryImm64) Type inferred from x.

UdivImm = 103

a = udiv_imm x, Y. (BinaryImm64) Type inferred from x.

SdivImm = 104

a = sdiv_imm x, Y. (BinaryImm64) Type inferred from x.

UremImm = 105

a = urem_imm x, Y. (BinaryImm64) Type inferred from x.

SremImm = 106

a = srem_imm x, Y. (BinaryImm64) Type inferred from x.

IrsubImm = 107

a = irsub_imm x, Y. (BinaryImm64) Type inferred from x.

IaddCin = 108

a = iadd_cin x, y, c_in. (Ternary) Type inferred from y.

IaddIfcin = 109

a = iadd_ifcin x, y, c_in. (Ternary) Type inferred from y.

IaddCout = 110

a, c_out = iadd_cout x, y. (Binary) Type inferred from x.

IaddIfcout = 111

a, c_out = iadd_ifcout x, y. (Binary) Type inferred from x.

IaddCarry = 112

a, c_out = iadd_carry x, y, c_in. (Ternary) Type inferred from y.

IaddIfcarry = 113

a, c_out = iadd_ifcarry x, y, c_in. (Ternary) Type inferred from y.

IsubBin = 114

a = isub_bin x, y, b_in. (Ternary) Type inferred from y.

IsubIfbin = 115

a = isub_ifbin x, y, b_in. (Ternary) Type inferred from y.

IsubBout = 116

a, b_out = isub_bout x, y. (Binary) Type inferred from x.

IsubIfbout = 117

a, b_out = isub_ifbout x, y. (Binary) Type inferred from x.

IsubBorrow = 118

a, b_out = isub_borrow x, y, b_in. (Ternary) Type inferred from y.

IsubIfborrow = 119

a, b_out = isub_ifborrow x, y, b_in. (Ternary) Type inferred from y.

Band = 120

a = band x, y. (Binary) Type inferred from x.

Bor = 121

a = bor x, y. (Binary) Type inferred from x.

Bxor = 122

a = bxor x, y. (Binary) Type inferred from x.

Bnot = 123

a = bnot x. (Unary) Type inferred from x.

BandNot = 124

a = band_not x, y. (Binary) Type inferred from x.

BorNot = 125

a = bor_not x, y. (Binary) Type inferred from x.

BxorNot = 126

a = bxor_not x, y. (Binary) Type inferred from x.

BandImm = 127

a = band_imm x, Y. (BinaryImm64) Type inferred from x.

BorImm = 128

a = bor_imm x, Y. (BinaryImm64) Type inferred from x.

BxorImm = 129

a = bxor_imm x, Y. (BinaryImm64) Type inferred from x.

Rotl = 130

a = rotl x, y. (Binary) Type inferred from x.

Rotr = 131

a = rotr x, y. (Binary) Type inferred from x.

RotlImm = 132

a = rotl_imm x, Y. (BinaryImm64) Type inferred from x.

RotrImm = 133

a = rotr_imm x, Y. (BinaryImm64) Type inferred from x.

Ishl = 134

a = ishl x, y. (Binary) Type inferred from x.

Ushr = 135

a = ushr x, y. (Binary) Type inferred from x.

Sshr = 136

a = sshr x, y. (Binary) Type inferred from x.

IshlImm = 137

a = ishl_imm x, Y. (BinaryImm64) Type inferred from x.

UshrImm = 138

a = ushr_imm x, Y. (BinaryImm64) Type inferred from x.

SshrImm = 139

a = sshr_imm x, Y. (BinaryImm64) Type inferred from x.

Bitrev = 140

a = bitrev x. (Unary) Type inferred from x.

Clz = 141

a = clz x. (Unary) Type inferred from x.

Cls = 142

a = cls x. (Unary) Type inferred from x.

Ctz = 143

a = ctz x. (Unary) Type inferred from x.

Popcnt = 144

a = popcnt x. (Unary) Type inferred from x.

Fcmp = 145

a = fcmp Cond, x, y. (FloatCompare) Type inferred from x.

Ffcmp = 146

f = ffcmp x, y. (Binary) Type inferred from x.

Fadd = 147

a = fadd x, y. (Binary) Type inferred from x.

Fsub = 148

a = fsub x, y. (Binary) Type inferred from x.

Fmul = 149

a = fmul x, y. (Binary) Type inferred from x.

Fdiv = 150

a = fdiv x, y. (Binary) Type inferred from x.

Sqrt = 151

a = sqrt x. (Unary) Type inferred from x.

Fma = 152

a = fma x, y, z. (Ternary) Type inferred from y.

Fneg = 153

a = fneg x. (Unary) Type inferred from x.

Fabs = 154

a = fabs x. (Unary) Type inferred from x.

Fcopysign = 155

a = fcopysign x, y. (Binary) Type inferred from x.

Fmin = 156

a = fmin x, y. (Binary) Type inferred from x.

FminPseudo = 157

a = fmin_pseudo x, y. (Binary) Type inferred from x.

Fmax = 158

a = fmax x, y. (Binary) Type inferred from x.

FmaxPseudo = 159

a = fmax_pseudo x, y. (Binary) Type inferred from x.

Ceil = 160

a = ceil x. (Unary) Type inferred from x.

Floor = 161

a = floor x. (Unary) Type inferred from x.

Trunc = 162

a = trunc x. (Unary) Type inferred from x.

Nearest = 163

a = nearest x. (Unary) Type inferred from x.

IsNull = 164

a = is_null x. (Unary) Type inferred from x.

IsInvalid = 165

a = is_invalid x. (Unary) Type inferred from x.

Trueif = 166

a = trueif Cond, f. (IntCond)

Trueff = 167

a = trueff Cond, f. (FloatCond)

Bitcast = 168

a = bitcast x. (Unary)

RawBitcast = 169

a = raw_bitcast x. (Unary)

ScalarToVector = 170

a = scalar_to_vector s. (Unary)

Breduce = 171

a = breduce x. (Unary)

Bextend = 172

a = bextend x. (Unary)

Bint = 173

a = bint x. (Unary)

Bmask = 174

a = bmask x. (Unary)

Ireduce = 175

a = ireduce x. (Unary)

Snarrow = 176

a = snarrow x, y. (Binary) Type inferred from x.

Unarrow = 177

a = unarrow x, y. (Binary) Type inferred from x.

Uunarrow = 178

a = uunarrow x, y. (Binary) Type inferred from x.

SwidenLow = 179

a = swiden_low x. (Unary) Type inferred from x.

SwidenHigh = 180

a = swiden_high x. (Unary) Type inferred from x.

UwidenLow = 181

a = uwiden_low x. (Unary) Type inferred from x.

UwidenHigh = 182

a = uwiden_high x. (Unary) Type inferred from x.

IaddPairwise = 183

a = iadd_pairwise x, y. (Binary) Type inferred from x.

WideningPairwiseDotProductS = 184

a = widening_pairwise_dot_product_s x, y. (Binary)

Uextend = 185

a = uextend x. (Unary)

Sextend = 186

a = sextend x. (Unary)

Fpromote = 187

a = fpromote x. (Unary)

Fdemote = 188

a = fdemote x. (Unary)

Fvdemote = 189

a = fvdemote x. (Unary)

FvpromoteLow = 190

x = fvpromote_low a. (Unary)

FcvtToUint = 191

a = fcvt_to_uint x. (Unary)

FcvtToSint = 192

a = fcvt_to_sint x. (Unary)

FcvtToUintSat = 193

a = fcvt_to_uint_sat x. (Unary)

FcvtToSintSat = 194

a = fcvt_to_sint_sat x. (Unary)

FcvtFromUint = 195

a = fcvt_from_uint x. (Unary)

FcvtFromSint = 196

a = fcvt_from_sint x. (Unary)

FcvtLowFromSint = 197

a = fcvt_low_from_sint x. (Unary)

Isplit = 198

lo, hi = isplit x. (Unary) Type inferred from x.

Iconcat = 199

a = iconcat lo, hi. (Binary) Type inferred from lo.

AtomicRmw = 200

a = atomic_rmw MemFlags, AtomicRmwOp, p, x. (AtomicRmw)

AtomicCas = 201

a = atomic_cas MemFlags, p, e, x. (AtomicCas) Type inferred from x.

AtomicLoad = 202

a = atomic_load MemFlags, p. (LoadNoOffset)

AtomicStore = 203

atomic_store MemFlags, x, p. (StoreNoOffset) Type inferred from x.

Fence = 204

fence. (NullAry)

ExtractVector = 205

a = extract_vector x, y. (BinaryImm8) Type inferred from x.

Implementations

impl Opcode

pub fn is_terminator(self) -> bool

True for instructions that terminate the block

pub fn is_branch(self) -> bool

True for all branch or jump instructions.

pub fn is_call(self) -> bool

Is this a call instruction?

pub fn is_return(self) -> bool

Is this a return instruction?

pub fn can_load(self) -> bool

Can this instruction read from memory?

pub fn can_store(self) -> bool

Can this instruction write to memory?

pub fn can_trap(self) -> bool

Can this instruction cause a trap?

pub fn other_side_effects(self) -> bool

Does this instruction have other side effects besides can_* flags?

pub fn writes_cpu_flags(self) -> bool

Does this instruction write to CPU flags?

impl Opcode

pub fn format(self) -> InstructionFormat

Get the instruction format for this opcode.

pub fn constraints(self) -> OpcodeConstraints

Get the constraint descriptor for this opcode. Panic if this is called on NotAnOpcode.

pub fn is_resumable_trap(&self) -> bool

Returns true if the instruction is a resumable trap.