cranelift_codegen::ir::instructions

Enum Opcode

#[repr(u8)]
pub enum Opcode {

    Jump = 1,
    Brif = 2,
    BrTable = 3,
    Debugtrap = 4,
    Trap = 5,
    Trapz = 6,
    ResumableTrap = 7,
    Trapnz = 8,
    ResumableTrapnz = 9,
    Return = 10,
    Call = 11,
    CallIndirect = 12,
    ReturnCall = 13,
    ReturnCallIndirect = 14,
    FuncAddr = 15,
    Splat = 16,
    Swizzle = 17,
    X86Pshufb = 18,
    Insertlane = 19,
    Extractlane = 20,
    Smin = 21,
    Umin = 22,
    Smax = 23,
    Umax = 24,
    AvgRound = 25,
    UaddSat = 26,
    SaddSat = 27,
    UsubSat = 28,
    SsubSat = 29,
    Load = 30,
    Store = 31,
    Uload8 = 32,
    Sload8 = 33,
    Istore8 = 34,
    Uload16 = 35,
    Sload16 = 36,
    Istore16 = 37,
    Uload32 = 38,
    Sload32 = 39,
    Istore32 = 40,
    Uload8x8 = 41,
    Sload8x8 = 42,
    Uload16x4 = 43,
    Sload16x4 = 44,
    Uload32x2 = 45,
    Sload32x2 = 46,
    StackLoad = 47,
    StackStore = 48,
    StackAddr = 49,
    DynamicStackLoad = 50,
    DynamicStackStore = 51,
    DynamicStackAddr = 52,
    GlobalValue = 53,
    SymbolValue = 54,
    TlsValue = 55,
    GetPinnedReg = 56,
    SetPinnedReg = 57,
    GetFramePointer = 58,
    GetStackPointer = 59,
    GetReturnAddress = 60,
    TableAddr = 61,
    Iconst = 62,
    F32const = 63,
    F64const = 64,
    Vconst = 65,
    Shuffle = 66,
    Null = 67,
    Nop = 68,
    Select = 69,
    SelectSpectreGuard = 70,
    Bitselect = 71,
    X86Blendv = 72,
    VanyTrue = 73,
    VallTrue = 74,
    VhighBits = 75,
    Icmp = 76,
    IcmpImm = 77,
    Iadd = 78,
    Isub = 79,
    Ineg = 80,
    Iabs = 81,
    Imul = 82,
    Umulhi = 83,
    Smulhi = 84,
    SqmulRoundSat = 85,
    X86Pmulhrsw = 86,
    Udiv = 87,
    Sdiv = 88,
    Urem = 89,
    Srem = 90,
    IaddImm = 91,
    ImulImm = 92,
    UdivImm = 93,
    SdivImm = 94,
    UremImm = 95,
    SremImm = 96,
    IrsubImm = 97,
    IaddCin = 98,
    IaddCout = 99,
    IaddCarry = 100,
    UaddOverflowTrap = 101,
    IsubBin = 102,
    IsubBout = 103,
    IsubBorrow = 104,
    Band = 105,
    Bor = 106,
    Bxor = 107,
    Bnot = 108,
    BandNot = 109,
    BorNot = 110,
    BxorNot = 111,
    BandImm = 112,
    BorImm = 113,
    BxorImm = 114,
    Rotl = 115,
    Rotr = 116,
    RotlImm = 117,
    RotrImm = 118,
    Ishl = 119,
    Ushr = 120,
    Sshr = 121,
    IshlImm = 122,
    UshrImm = 123,
    SshrImm = 124,
    Bitrev = 125,
    Clz = 126,
    Cls = 127,
    Ctz = 128,
    Bswap = 129,
    Popcnt = 130,
    Fcmp = 131,
    Fadd = 132,
    Fsub = 133,
    Fmul = 134,
    Fdiv = 135,
    Sqrt = 136,
    Fma = 137,
    Fneg = 138,
    Fabs = 139,
    Fcopysign = 140,
    Fmin = 141,
    FminPseudo = 142,
    Fmax = 143,
    FmaxPseudo = 144,
    Ceil = 145,
    Floor = 146,
    Trunc = 147,
    Nearest = 148,
    IsNull = 149,
    IsInvalid = 150,
    Bitcast = 151,
    ScalarToVector = 152,
    Bmask = 153,
    Ireduce = 154,
    Snarrow = 155,
    Unarrow = 156,
    Uunarrow = 157,
    SwidenLow = 158,
    SwidenHigh = 159,
    UwidenLow = 160,
    UwidenHigh = 161,
    IaddPairwise = 162,
    X86Pmaddubsw = 163,
    Uextend = 164,
    Sextend = 165,
    Fpromote = 166,
    Fdemote = 167,
    Fvdemote = 168,
    FvpromoteLow = 169,
    FcvtToUint = 170,
    FcvtToSint = 171,
    FcvtToUintSat = 172,
    FcvtToSintSat = 173,
    X86Cvtt2dq = 174,
    FcvtFromUint = 175,
    FcvtFromSint = 176,
    FcvtLowFromSint = 177,
    Isplit = 178,
    Iconcat = 179,
    AtomicRmw = 180,
    AtomicCas = 181,
    AtomicLoad = 182,
    AtomicStore = 183,
    Fence = 184,
    ExtractVector = 185,

}

An instruction opcode.

All instructions from all supported ISAs are present.

Variants

Jump = 1

jump block_call. (Jump)

Brif = 2

brif c, block_then, block_else. (Brif) Type inferred from c.

BrTable = 3

br_table x, JT. (BranchTable)

Debugtrap = 4

debugtrap. (NullAry)

Trap = 5

trap code. (Trap)

Trapz = 6

trapz c, code. (CondTrap) Type inferred from c.

ResumableTrap = 7

resumable_trap code. (Trap)

Trapnz = 8

trapnz c, code. (CondTrap) Type inferred from c.

ResumableTrapnz = 9

resumable_trapnz c, code. (CondTrap) Type inferred from c.

Return = 10

return rvals. (MultiAry)

Call = 11

rvals = call FN, args. (Call)

CallIndirect = 12

rvals = call_indirect SIG, callee, args. (CallIndirect) Type inferred from callee.

ReturnCall = 13

return_call FN, args. (Call)

ReturnCallIndirect = 14

return_call_indirect SIG, callee, args. (CallIndirect) Type inferred from callee.

FuncAddr = 15

addr = func_addr FN. (FuncAddr)

Splat = 16

a = splat x. (Unary)

Swizzle = 17

a = swizzle x, y. (Binary)

X86Pshufb = 18

a = x86_pshufb x, y. (Binary)

Insertlane = 19

a = insertlane x, y, Idx. (TernaryImm8) Type inferred from x.

Extractlane = 20

a = extractlane x, Idx. (BinaryImm8) Type inferred from x.

Smin = 21

a = smin x, y. (Binary) Type inferred from x.

Umin = 22

a = umin x, y. (Binary) Type inferred from x.

Smax = 23

a = smax x, y. (Binary) Type inferred from x.

Umax = 24

a = umax x, y. (Binary) Type inferred from x.

AvgRound = 25

a = avg_round x, y. (Binary) Type inferred from x.

UaddSat = 26

a = uadd_sat x, y. (Binary) Type inferred from x.

SaddSat = 27

a = sadd_sat x, y. (Binary) Type inferred from x.

UsubSat = 28

a = usub_sat x, y. (Binary) Type inferred from x.

SsubSat = 29

a = ssub_sat x, y. (Binary) Type inferred from x.

Load = 30

a = load MemFlags, p, Offset. (Load)

Store = 31

store MemFlags, x, p, Offset. (Store) Type inferred from x.

Uload8 = 32

a = uload8 MemFlags, p, Offset. (Load)

Sload8 = 33

a = sload8 MemFlags, p, Offset. (Load)

Istore8 = 34

istore8 MemFlags, x, p, Offset. (Store) Type inferred from x.

Uload16 = 35

a = uload16 MemFlags, p, Offset. (Load)

Sload16 = 36

a = sload16 MemFlags, p, Offset. (Load)

Istore16 = 37

istore16 MemFlags, x, p, Offset. (Store) Type inferred from x.

Uload32 = 38

a = uload32 MemFlags, p, Offset. (Load) Type inferred from p.

Sload32 = 39

a = sload32 MemFlags, p, Offset. (Load) Type inferred from p.

Istore32 = 40

istore32 MemFlags, x, p, Offset. (Store) Type inferred from x.

Uload8x8 = 41

a = uload8x8 MemFlags, p, Offset. (Load) Type inferred from p.

Sload8x8 = 42

a = sload8x8 MemFlags, p, Offset. (Load) Type inferred from p.

Uload16x4 = 43

a = uload16x4 MemFlags, p, Offset. (Load) Type inferred from p.

Sload16x4 = 44

a = sload16x4 MemFlags, p, Offset. (Load) Type inferred from p.

Uload32x2 = 45

a = uload32x2 MemFlags, p, Offset. (Load) Type inferred from p.

Sload32x2 = 46

a = sload32x2 MemFlags, p, Offset. (Load) Type inferred from p.

StackLoad = 47

a = stack_load SS, Offset. (StackLoad)

StackStore = 48

stack_store x, SS, Offset. (StackStore) Type inferred from x.

StackAddr = 49

addr = stack_addr SS, Offset. (StackLoad)

DynamicStackLoad = 50

a = dynamic_stack_load DSS. (DynamicStackLoad)

DynamicStackStore = 51

dynamic_stack_store x, DSS. (DynamicStackStore) Type inferred from x.

DynamicStackAddr = 52

addr = dynamic_stack_addr DSS. (DynamicStackLoad)

GlobalValue = 53

a = global_value GV. (UnaryGlobalValue)

SymbolValue = 54

a = symbol_value GV. (UnaryGlobalValue)

TlsValue = 55

a = tls_value GV. (UnaryGlobalValue)

GetPinnedReg = 56

addr = get_pinned_reg. (NullAry)

SetPinnedReg = 57

set_pinned_reg addr. (Unary) Type inferred from addr.

GetFramePointer = 58

addr = get_frame_pointer. (NullAry)

GetStackPointer = 59

addr = get_stack_pointer. (NullAry)

GetReturnAddress = 60

addr = get_return_address. (NullAry)

TableAddr = 61

addr = table_addr T, p, Offset. (TableAddr)

Iconst = 62

a = iconst N. (UnaryImm)

F32const = 63

a = f32const N. (UnaryIeee32)

F64const = 64

a = f64const N. (UnaryIeee64)

Vconst = 65

a = vconst N. (UnaryConst)

Shuffle = 66

a = shuffle a, b, mask. (Shuffle)

Null = 67

a = null. (NullAry)

Nop = 68

nop. (NullAry)

Select = 69

a = select c, x, y. (Ternary) Type inferred from x.

SelectSpectreGuard = 70

a = select_spectre_guard c, x, y. (Ternary) Type inferred from x.

Bitselect = 71

a = bitselect c, x, y. (Ternary) Type inferred from x.

X86Blendv = 72

a = x86_blendv c, x, y. (Ternary) Type inferred from x.

VanyTrue = 73

s = vany_true a. (Unary) Type inferred from a.

VallTrue = 74

s = vall_true a. (Unary) Type inferred from a.

VhighBits = 75

x = vhigh_bits a. (Unary)

Icmp = 76

a = icmp Cond, x, y. (IntCompare) Type inferred from x.

IcmpImm = 77

a = icmp_imm Cond, x, Y. (IntCompareImm) Type inferred from x.

Iadd = 78

a = iadd x, y. (Binary) Type inferred from x.

Isub = 79

a = isub x, y. (Binary) Type inferred from x.

Ineg = 80

a = ineg x. (Unary) Type inferred from x.

Iabs = 81

a = iabs x. (Unary) Type inferred from x.

Imul = 82

a = imul x, y. (Binary) Type inferred from x.

Umulhi = 83

a = umulhi x, y. (Binary) Type inferred from x.

Smulhi = 84

a = smulhi x, y. (Binary) Type inferred from x.

SqmulRoundSat = 85

a = sqmul_round_sat x, y. (Binary) Type inferred from x.

X86Pmulhrsw = 86

a = x86_pmulhrsw x, y. (Binary) Type inferred from x.

Udiv = 87

a = udiv x, y. (Binary) Type inferred from x.

Sdiv = 88

a = sdiv x, y. (Binary) Type inferred from x.

Urem = 89

a = urem x, y. (Binary) Type inferred from x.

Srem = 90

a = srem x, y. (Binary) Type inferred from x.

IaddImm = 91

a = iadd_imm x, Y. (BinaryImm64) Type inferred from x.

ImulImm = 92

a = imul_imm x, Y. (BinaryImm64) Type inferred from x.

UdivImm = 93

a = udiv_imm x, Y. (BinaryImm64) Type inferred from x.

SdivImm = 94

a = sdiv_imm x, Y. (BinaryImm64) Type inferred from x.

UremImm = 95

a = urem_imm x, Y. (BinaryImm64) Type inferred from x.

SremImm = 96

a = srem_imm x, Y. (BinaryImm64) Type inferred from x.

IrsubImm = 97

a = irsub_imm x, Y. (BinaryImm64) Type inferred from x.

IaddCin = 98

a = iadd_cin x, y, c_in. (Ternary) Type inferred from y.

IaddCout = 99

a, c_out = iadd_cout x, y. (Binary) Type inferred from x.

IaddCarry = 100

a, c_out = iadd_carry x, y, c_in. (Ternary) Type inferred from y.

UaddOverflowTrap = 101

a = uadd_overflow_trap x, y, code. (IntAddTrap) Type inferred from x.

IsubBin = 102

a = isub_bin x, y, b_in. (Ternary) Type inferred from y.

IsubBout = 103

a, b_out = isub_bout x, y. (Binary) Type inferred from x.

IsubBorrow = 104

a, b_out = isub_borrow x, y, b_in. (Ternary) Type inferred from y.

Band = 105

a = band x, y. (Binary) Type inferred from x.

Bor = 106

a = bor x, y. (Binary) Type inferred from x.

Bxor = 107

a = bxor x, y. (Binary) Type inferred from x.

Bnot = 108

a = bnot x. (Unary) Type inferred from x.

BandNot = 109

a = band_not x, y. (Binary) Type inferred from x.

BorNot = 110

a = bor_not x, y. (Binary) Type inferred from x.

BxorNot = 111

a = bxor_not x, y. (Binary) Type inferred from x.

BandImm = 112

a = band_imm x, Y. (BinaryImm64) Type inferred from x.

BorImm = 113

a = bor_imm x, Y. (BinaryImm64) Type inferred from x.

BxorImm = 114

a = bxor_imm x, Y. (BinaryImm64) Type inferred from x.

Rotl = 115

a = rotl x, y. (Binary) Type inferred from x.

Rotr = 116

a = rotr x, y. (Binary) Type inferred from x.

RotlImm = 117

a = rotl_imm x, Y. (BinaryImm64) Type inferred from x.

RotrImm = 118

a = rotr_imm x, Y. (BinaryImm64) Type inferred from x.

Ishl = 119

a = ishl x, y. (Binary) Type inferred from x.

Ushr = 120

a = ushr x, y. (Binary) Type inferred from x.

Sshr = 121

a = sshr x, y. (Binary) Type inferred from x.

IshlImm = 122

a = ishl_imm x, Y. (BinaryImm64) Type inferred from x.

UshrImm = 123

a = ushr_imm x, Y. (BinaryImm64) Type inferred from x.

SshrImm = 124

a = sshr_imm x, Y. (BinaryImm64) Type inferred from x.

Bitrev = 125

a = bitrev x. (Unary) Type inferred from x.

Clz = 126

a = clz x. (Unary) Type inferred from x.

Cls = 127

a = cls x. (Unary) Type inferred from x.

Ctz = 128

a = ctz x. (Unary) Type inferred from x.

Bswap = 129

a = bswap x. (Unary) Type inferred from x.

Popcnt = 130

a = popcnt x. (Unary) Type inferred from x.

Fcmp = 131

a = fcmp Cond, x, y. (FloatCompare) Type inferred from x.

Fadd = 132

a = fadd x, y. (Binary) Type inferred from x.

Fsub = 133

a = fsub x, y. (Binary) Type inferred from x.

Fmul = 134

a = fmul x, y. (Binary) Type inferred from x.

Fdiv = 135

a = fdiv x, y. (Binary) Type inferred from x.

Sqrt = 136

a = sqrt x. (Unary) Type inferred from x.

Fma = 137

a = fma x, y, z. (Ternary) Type inferred from y.

Fneg = 138

a = fneg x. (Unary) Type inferred from x.

Fabs = 139

a = fabs x. (Unary) Type inferred from x.

Fcopysign = 140

a = fcopysign x, y. (Binary) Type inferred from x.

Fmin = 141

a = fmin x, y. (Binary) Type inferred from x.

FminPseudo = 142

a = fmin_pseudo x, y. (Binary) Type inferred from x.

Fmax = 143

a = fmax x, y. (Binary) Type inferred from x.

FmaxPseudo = 144

a = fmax_pseudo x, y. (Binary) Type inferred from x.

Ceil = 145

a = ceil x. (Unary) Type inferred from x.

Floor = 146

a = floor x. (Unary) Type inferred from x.

Trunc = 147

a = trunc x. (Unary) Type inferred from x.

Nearest = 148

a = nearest x. (Unary) Type inferred from x.

IsNull = 149

a = is_null x. (Unary) Type inferred from x.

IsInvalid = 150

a = is_invalid x. (Unary) Type inferred from x.

Bitcast = 151

a = bitcast MemFlags, x. (LoadNoOffset)

ScalarToVector = 152

a = scalar_to_vector s. (Unary)

Bmask = 153

a = bmask x. (Unary)

Ireduce = 154

a = ireduce x. (Unary)

Snarrow = 155

a = snarrow x, y. (Binary) Type inferred from x.

Unarrow = 156

a = unarrow x, y. (Binary) Type inferred from x.

Uunarrow = 157

a = uunarrow x, y. (Binary) Type inferred from x.

SwidenLow = 158

a = swiden_low x. (Unary) Type inferred from x.

SwidenHigh = 159

a = swiden_high x. (Unary) Type inferred from x.

UwidenLow = 160

a = uwiden_low x. (Unary) Type inferred from x.

UwidenHigh = 161

a = uwiden_high x. (Unary) Type inferred from x.

IaddPairwise = 162

a = iadd_pairwise x, y. (Binary) Type inferred from x.

X86Pmaddubsw = 163

a = x86_pmaddubsw x, y. (Binary)

Uextend = 164

a = uextend x. (Unary)

Sextend = 165

a = sextend x. (Unary)

Fpromote = 166

a = fpromote x. (Unary)

Fdemote = 167

a = fdemote x. (Unary)

Fvdemote = 168

a = fvdemote x. (Unary)

FvpromoteLow = 169

x = fvpromote_low a. (Unary)

FcvtToUint = 170

a = fcvt_to_uint x. (Unary)

FcvtToSint = 171

a = fcvt_to_sint x. (Unary)

FcvtToUintSat = 172

a = fcvt_to_uint_sat x. (Unary)

FcvtToSintSat = 173

a = fcvt_to_sint_sat x. (Unary)

X86Cvtt2dq = 174

a = x86_cvtt2dq x. (Unary)

FcvtFromUint = 175

a = fcvt_from_uint x. (Unary)

FcvtFromSint = 176

a = fcvt_from_sint x. (Unary)

FcvtLowFromSint = 177

a = fcvt_low_from_sint x. (Unary)

Isplit = 178

lo, hi = isplit x. (Unary) Type inferred from x.

Iconcat = 179

a = iconcat lo, hi. (Binary) Type inferred from lo.

AtomicRmw = 180

a = atomic_rmw MemFlags, AtomicRmwOp, p, x. (AtomicRmw)

AtomicCas = 181

a = atomic_cas MemFlags, p, e, x. (AtomicCas) Type inferred from x.

AtomicLoad = 182

a = atomic_load MemFlags, p. (LoadNoOffset)

AtomicStore = 183

atomic_store MemFlags, x, p. (StoreNoOffset) Type inferred from x.

Fence = 184

fence. (NullAry)

ExtractVector = 185

a = extract_vector x, y. (BinaryImm8) Type inferred from x.

Implementations

impl Opcode

pub fn is_terminator(self) -> bool

True for instructions that terminate the block

pub fn is_branch(self) -> bool

True for all branch or jump instructions.

pub fn is_call(self) -> bool

Is this a call instruction?

pub fn is_return(self) -> bool

Is this a return instruction?

pub fn can_load(self) -> bool

Can this instruction read from memory?

pub fn can_store(self) -> bool

Can this instruction write to memory?

pub fn can_trap(self) -> bool

Can this instruction cause a trap?

pub fn other_side_effects(self) -> bool

Does this instruction have other side effects besides can_* flags?

pub fn side_effects_idempotent(self) -> bool

Despite having side effects, is this instruction okay to GVN?

pub fn all() -> &'static [Opcode]

All cranelift opcodes.

impl Opcode

pub fn format(self) -> InstructionFormat

Get the instruction format for this opcode.

pub fn constraints(self) -> OpcodeConstraints

Get the constraint descriptor for this opcode. Panic if this is called on NotAnOpcode.

pub fn is_resumable_trap(&self) -> bool

Returns true if the instruction is a resumable trap.