Module cranelift_codegen::ir::layout

Function layout.

The order of basic blocks in a function and the order of instructions in a block is determined by the Layout data structure defined in this module.

Structs

Blocks

Iterate over blocks in layout order. See Layout::blocks().

Insts

Iterate over instructions in a block in layout order. See Layout::block_insts().

Layout

The Layout struct determines the layout of blocks and instructions in a function. It does not contain definitions of instructions or blocks, but depends on Inst and Block entity references being defined elsewhere.