Module cranelift_codegen::ir::layout

Function layout.

The order of extended basic blocks in a function and the order of instructions in an EBB is determined by the Layout data structure defined in this module.

Structs

Ebbs
Iterate over EBBs in layout order. See Layout::ebbs().

Insts
Iterate over instructions in an EBB in layout order. See Layout::ebb_insts().

Layout
The Layout struct determines the layout of EBBs and instructions in a function. It does not contain definitions of instructions or EBBs, but depends on Inst and Ebb entity references being defined elsewhere.