Struct cranelift_codegen::ir::layout::Layout

pub struct Layout { /* fields omitted */ }

The Layout struct determines the layout of blocks and instructions in a function. It does not contain definitions of instructions or blocks, but depends on Inst and Block entity references being defined elsewhere.

This data structure determines:

The order of blocks in the function.

Which block contains a given instruction.

The order of instructions with a block.

While data dependencies are not recorded, instruction ordering does affect control dependencies, so part of the semantics of the program are determined by the layout.

Implementations

impl Layout

pub fn new() -> Self

Create a new empty Layout.

pub fn clear(&mut self)

Clear the layout.

pub fn block_capacity(&self) -> usize

Returns the capacity of the BlockData map.

impl Layout

Methods for laying out blocks.

An unknown block starts out as not inserted in the block layout. The layout is a linear order of inserted blocks. Once a block has been inserted in the layout, instructions can be added. A block can only be removed from the layout when it is empty.

Since every block must end with a terminator instruction which cannot fall through, the layout of blocks do not affect the semantics of the program.

pub fn is_block_inserted(&self, block: Block) -> bool

Is block currently part of the layout?

pub fn append_block(&mut self, block: Block)

Insert block as the last block in the layout.

pub fn insert_block(&mut self, block: Block, before: Block)

Insert block in the layout before the existing block before.

pub fn insert_block_after(&mut self, block: Block, after: Block)

Insert block in the layout after the existing block after.

pub fn remove_block(&mut self, block: Block)

Remove block from the layout.

pub fn blocks(&self) -> Blocks<'_>

Return an iterator over all blocks in layout order.

pub fn entry_block(&self) -> Option<Block>

Get the function’s entry block. This is simply the first block in the layout order.

pub fn last_block(&self) -> Option<Block>

Get the last block in the layout.

pub fn prev_block(&self, block: Block) -> Option<Block>

Get the block preceding block in the layout order.

pub fn next_block(&self, block: Block) -> Option<Block>

Get the block following block in the layout order.

impl Layout

Methods for arranging instructions.

An instruction starts out as not inserted in the layout. An instruction can be inserted into a block at a given position.

pub fn inst_block(&self, inst: Inst) -> Option<Block>

Get the block containing inst, or None if inst is not inserted in the layout.

pub fn pp_block<PP>(&self, pp: PP) -> Block where
PP: Into<ExpandedProgramPoint>,

Get the block containing the program point pp. Panic if pp is not in the layout.

pub fn append_inst(&mut self, inst: Inst, block: Block)

Append inst to the end of block.

pub fn first_inst(&self, block: Block) -> Option<Inst>

Fetch a block’s first instruction.

pub fn last_inst(&self, block: Block) -> Option<Inst>

Fetch a block’s last instruction.

pub fn next_inst(&self, inst: Inst) -> Option<Inst>

Fetch the instruction following inst.

pub fn prev_inst(&self, inst: Inst) -> Option<Inst>

Fetch the instruction preceding inst.

pub fn canonical_branch_inst(
&self,
dfg: &DataFlowGraph,
block: Block
) -> Option<Inst>

Fetch the first instruction in a block’s terminal branch group.

pub fn insert_inst(&mut self, inst: Inst, before: Inst)

Insert inst before the instruction before in the same block.

pub fn remove_inst(&mut self, inst: Inst)

Remove inst from the layout.

pub fn block_insts(&self, block: Block) -> Insts<'_>

Iterate over the instructions in block in layout order.

pub fn block_likely_branches(&self, block: Block) -> Insts<'_>

Iterate over a limited set of instruction which are likely the branches of block in layout order. Any instruction not visited by this iterator is not a branch, but an instruction visited by this may not be a branch.

pub fn split_block(&mut self, new_block: Block, before: Inst)

Split the block containing before in two.

Insert new_block after the old block and move before and the following instructions to new_block:

old_block:
    i1
    i2
    i3 << before
    i4

becomes:

old_block:
    i1
    i2
new_block:
    i3 << before
    i4