Struct cranelift_codegen::ir::Signature

pub struct Signature {
    pub params: Vec<AbiParam>,
    pub returns: Vec<AbiParam>,
    pub call_conv: CallConv,
}

Function signature.

The function signature describes the types of formal parameters and return values along with other details that are needed to call a function correctly.

A signature can optionally include ISA-specific ABI information which specifies exactly how arguments and return values are passed.

Fields
params: Vec<AbiParam>

The arguments passed to the function.

returns: Vec<AbiParam>

Values returned from the function.

call_conv: CallConv

Calling convention.

Methods

impl Signature

pub fn new(call_conv: CallConv) -> Self

Create a new blank signature.

pub fn clear(&mut self, call_conv: CallConv)

Clear the signature so it is identical to a fresh one returned by new().

pub fn display<'a, R: Into<Option<&'a RegInfo>>>(
&'a self,
regs: R
) -> DisplaySignature<'a>

Return an object that can display self with correct register names.

pub fn special_param_index(&self, purpose: ArgumentPurpose) -> Option<usize>

Find the index of a presumed unique special-purpose parameter.

pub fn special_return_index(&self, purpose: ArgumentPurpose) -> Option<usize>

Find the index of a presumed unique special-purpose parameter.

pub fn uses_special_param(&self, purpose: ArgumentPurpose) -> bool

Does this signature have a parameter whose ArgumentPurpose is purpose?

pub fn uses_special_return(&self, purpose: ArgumentPurpose) -> bool

Does this signature have a return whose ArgumentPurpose is purpose?

pub fn num_special_params(&self) -> usize

How many special parameters does this function have?

pub fn num_special_returns(&self) -> usize

How many special returns does this function have?

pub fn uses_struct_return_param(&self) -> bool

Does this signature take an struct return pointer parameter?

pub fn is_multi_return(&self) -> bool

Does this return more than one normal value? (Pre-struct return legalization)