Struct cranelift_codegen::ir::types::Type

pub struct Type(_);

The type of an SSA value.

The INVALID type isn't a real type, and is used as a placeholder in the IR where a type field is present put no type is needed, such as the controlling type variable for a non-polymorphic instruction.

Basic integer types: I8, I16, I32, I64, and I128. These types are sign-agnostic.

Basic floating point types: F32 and F64. IEEE single and double precision.

Boolean types: B1, B8, B16, B32, B64, and B128. These all encode 'true' or 'false'. The larger types use redundant bits.

SIMD vector types have power-of-two lanes, up to 256. Lanes can be any int/float/bool type.

Methods

impl Type

pub fn lane_type(self) -> Self

Get the lane type of this SIMD vector type.

A lane type is the same as a SIMD vector type with one lane, so it returns itself.

pub fn lane_of(self) -> Self

The type transformation that returns the lane type of a type variable; it is just a renaming of lane_type() to be used in context where we think in terms of type variable transformations.

pub fn log2_lane_bits(self) -> u8

Get log_2 of the number of bits in a lane.

pub fn lane_bits(self) -> u8

Get the number of bits in a lane.

pub fn int(bits: u16) -> Option<Self>

Get an integer type with the requested number of bits.

pub fn as_bool_pedantic(self) -> Self

Get a type with the same number of lanes as this type, but with the lanes replaced by booleans of the same size.

Lane types are treated as vectors with one lane, so they are converted to the multi-bit boolean types.

pub fn as_bool(self) -> Self

Get a type with the same number of lanes as this type, but with the lanes replaced by booleans of the same size.

Scalar types are all converted to b1 which is usually what you want.

pub fn half_width(self) -> Option<Self>

Get a type with the same number of lanes as this type, but with lanes that are half the number of bits.

pub fn double_width(self) -> Option<Self>

Get a type with the same number of lanes as this type, but with lanes that are twice the number of bits.

pub fn is_invalid(self) -> bool

Is this the INVALID type?

pub fn is_special(self) -> bool

Is this a special type?

pub fn is_lane(self) -> bool

Is this a lane type?

This is a scalar type that can also appear as the lane type of a SIMD vector.

pub fn is_vector(self) -> bool

Is this a SIMD vector type?

A vector type has 2 or more lanes.

pub fn is_bool(self) -> bool

Is this a scalar boolean type?

pub fn is_int(self) -> bool

Is this a scalar integer type?

pub fn is_float(self) -> bool

Is this a scalar floating point type?

pub fn is_flags(self) -> bool

Is this a CPU flags type?

pub fn is_ref(self) -> bool

Is this a ref type?

pub fn log2_lane_count(self) -> u8

Get log_2 of the number of lanes in this SIMD vector type.

All SIMD types have a lane count that is a power of two and no larger than 256, so this will be a number in the range 0-8.

A scalar type is the same as a SIMD vector type with one lane, so it returns 0.

pub fn lane_count(self) -> u16

Get the number of lanes in this SIMD vector type.

A scalar type is the same as a SIMD vector type with one lane, so it returns 1.

pub fn bits(self) -> u16

Get the total number of bits used to represent this type.

pub fn bytes(self) -> u32

Get the number of bytes used to store this type in memory.

pub fn by(self, n: u16) -> Option<Self>

Get a SIMD vector type with n times more lanes than this one.

If this is a scalar type, this produces a SIMD type with this as a lane type and n lanes.

If this is already a SIMD vector type, this produces a SIMD vector type with n * self.lane_count() lanes.

pub fn half_vector(self) -> Option<Self>

Get a SIMD vector with half the number of lanes.

There is no double_vector() method. Use t.by(2) instead.

pub fn index(self) -> usize

Index of this type, for use with hash tables etc.

pub fn wider_or_equal(self, other: Self) -> bool

True iff:

self.lane_count() == other.lane_count() and

self.lane_bits() >= other.lane_bits()

pub fn triple_pointer_type(triple: &Triple) -> Self

Return the pointer type for the given target triple.