Module cranelift_codegen::isa::aarch64::inst

This module defines aarch64-specific machine instruction types.

Re-exports

pub use self::imms::*;

pub use self::args::*;

Modules

args

AArch64 ISA definitions: instruction arguments.

emit

AArch64 ISA: binary code emission.

imms

AArch64 ISA definitions: immediate constants.

Structs

CallIndInfo

Additional information for CallInd instructions, left out of line to lower the size of the Inst enum.

CallInfo

Additional information for (direct) Call instructions, left out of line to lower the size of the Inst enum.

ReturnCallInfo

Additional information for return_call[_ind] instructions, left out of line to lower the size of the Inst enum.

Enums

ALUOp

Internal type ALUOp: defined at src/isa/aarch64/inst.isle line 986.

ALUOp3

Internal type ALUOp3: defined at src/isa/aarch64/inst.isle line 1024.

AMode

Internal type AMode: defined at src/isa/aarch64/inst.isle line 1093.

APIKey

Internal type APIKey: defined at src/isa/aarch64/inst.isle line 1673.

AtomicRMWLoopOp

Internal type AtomicRMWLoopOp: defined at src/isa/aarch64/inst.isle line 1657.

AtomicRMWOp

Internal type AtomicRMWOp: defined at src/isa/aarch64/inst.isle line 1642.

BitOp

Internal type BitOp: defined at src/isa/aarch64/inst.isle line 1076.

BranchTargetType

Internal type BranchTargetType: defined at src/isa/aarch64/inst.isle line 1686.

FPUOp1

Internal type FPUOp1: defined at src/isa/aarch64/inst.isle line 1311.

FPUOp2

Internal type FPUOp2: defined at src/isa/aarch64/inst.isle line 1321.

FPUOp3

Internal type FPUOp3: defined at src/isa/aarch64/inst.isle line 1332.

FPUOpRI

A floating-point unit (FPU) operation with two args, a register and an immediate.

FPUOpRIMod

A floating-point unit (FPU) operation with two args, a register and an immediate that modifies its dest (so takes that input value as a separate virtual register).

FpuRoundMode

Internal type FpuRoundMode: defined at src/isa/aarch64/inst.isle line 1365.

FpuToIntOp

Internal type FpuToIntOp: defined at src/isa/aarch64/inst.isle line 1338.

Inst

Internal type MInst: defined at src/isa/aarch64/inst.isle line 2.

IntToFpuOp

Internal type IntToFpuOp: defined at src/isa/aarch64/inst.isle line 1351.

LabelUse

Different forms of label references for different instruction formats.

MoveWideOp

Internal type MoveWideOp: defined at src/isa/aarch64/inst.isle line 1036.

VecALUModOp

Internal type VecALUModOp: defined at src/isa/aarch64/inst.isle line 1474.

VecALUOp

Internal type VecALUOp: defined at src/isa/aarch64/inst.isle line 1387.

VecExtendOp

Internal type VecExtendOp: defined at src/isa/aarch64/inst.isle line 1378.

VecLanesOp

Internal type VecLanesOp: defined at src/isa/aarch64/inst.isle line 1615.

VecMisc2

Internal type VecMisc2: defined at src/isa/aarch64/inst.isle line 1485.

VecPairOp

Internal type VecPairOp: defined at src/isa/aarch64/inst.isle line 1596.

VecRRLongOp

Internal type VecRRLongOp: defined at src/isa/aarch64/inst.isle line 1546.

VecRRNarrowOp

Internal type VecRRNarrowOp: defined at src/isa/aarch64/inst.isle line 1561.

VecRRPairLongOp

Internal type VecRRPairLongOp: defined at src/isa/aarch64/inst.isle line 1604.

VecRRRLongModOp

Internal type VecRRRLongModOp: defined at src/isa/aarch64/inst.isle line 1587.

VecRRRLongOp

Internal type VecRRRLongOp: defined at src/isa/aarch64/inst.isle line 1575.

VecShiftImmModOp

Internal type VecShiftImmModOp: defined at src/isa/aarch64/inst.isle line 1635.

VecShiftImmOp

Internal type VecShiftImmOp: defined at src/isa/aarch64/inst.isle line 1624.