Module cranelift_codegen::isa::aarch64::inst

This module defines aarch64-specific machine instruction types.

Re-exports

pub use self::imms::*;

pub use self::args::*;

Modules

args

AArch64 ISA definitions: instruction arguments.

emit

AArch64 ISA: binary code emission.

imms

AArch64 ISA definitions: immediate constants.

Structs

CallIndInfo

Additional information for CallInd instructions, left out of line to lower the size of the Inst enum.

CallInfo

Additional information for (direct) Call instructions, left out of line to lower the size of the Inst enum.

JTSequenceInfo

Additional information for JTSequence instructions, left out of line to lower the size of the Inst enum.

ReturnCallInfo

Additional information for return_call[_ind] instructions, left out of line to lower the size of the Inst enum.

Enums

ALUOp

Internal type ALUOp: defined at src/isa/aarch64/inst.isle line 975.

ALUOp3

Internal type ALUOp3: defined at src/isa/aarch64/inst.isle line 1013.

AMode

Internal type AMode: defined at src/isa/aarch64/inst.isle line 1082.

APIKey

Internal type APIKey: defined at src/isa/aarch64/inst.isle line 1660.

AtomicRMWLoopOp

Internal type AtomicRMWLoopOp: defined at src/isa/aarch64/inst.isle line 1644.

AtomicRMWOp

Internal type AtomicRMWOp: defined at src/isa/aarch64/inst.isle line 1629.

BitOp

Internal type BitOp: defined at src/isa/aarch64/inst.isle line 1065.

BranchTargetType

Internal type BranchTargetType: defined at src/isa/aarch64/inst.isle line 1673.

FPUOp1

Internal type FPUOp1: defined at src/isa/aarch64/inst.isle line 1298.

FPUOp2

Internal type FPUOp2: defined at src/isa/aarch64/inst.isle line 1308.

FPUOp3

Internal type FPUOp3: defined at src/isa/aarch64/inst.isle line 1319.

FPUOpRI

A floating-point unit (FPU) operation with two args, a register and an immediate.

FPUOpRIMod

A floating-point unit (FPU) operation with two args, a register and an immediate that modifies its dest (so takes that input value as a separate virtual register).

FpuRoundMode

Internal type FpuRoundMode: defined at src/isa/aarch64/inst.isle line 1352.

FpuToIntOp

Internal type FpuToIntOp: defined at src/isa/aarch64/inst.isle line 1325.

Inst

Internal type MInst: defined at src/isa/aarch64/inst.isle line 2.

IntToFpuOp

Internal type IntToFpuOp: defined at src/isa/aarch64/inst.isle line 1338.

LabelUse

Different forms of label references for different instruction formats.

MoveWideOp

Internal type MoveWideOp: defined at src/isa/aarch64/inst.isle line 1025.

VecALUModOp

Internal type VecALUModOp: defined at src/isa/aarch64/inst.isle line 1461.

VecALUOp

Internal type VecALUOp: defined at src/isa/aarch64/inst.isle line 1374.

VecExtendOp

Internal type VecExtendOp: defined at src/isa/aarch64/inst.isle line 1365.

VecLanesOp

Internal type VecLanesOp: defined at src/isa/aarch64/inst.isle line 1602.

VecMisc2

Internal type VecMisc2: defined at src/isa/aarch64/inst.isle line 1472.

VecPairOp

Internal type VecPairOp: defined at src/isa/aarch64/inst.isle line 1583.

VecRRLongOp

Internal type VecRRLongOp: defined at src/isa/aarch64/inst.isle line 1533.

VecRRNarrowOp

Internal type VecRRNarrowOp: defined at src/isa/aarch64/inst.isle line 1548.

VecRRPairLongOp

Internal type VecRRPairLongOp: defined at src/isa/aarch64/inst.isle line 1591.

VecRRRLongModOp

Internal type VecRRRLongModOp: defined at src/isa/aarch64/inst.isle line 1574.

VecRRRLongOp

Internal type VecRRRLongOp: defined at src/isa/aarch64/inst.isle line 1562.

VecShiftImmModOp

Internal type VecShiftImmModOp: defined at src/isa/aarch64/inst.isle line 1622.

VecShiftImmOp

Internal type VecShiftImmOp: defined at src/isa/aarch64/inst.isle line 1611.