Trait cranelift_codegen::isa::TargetIsa

pub trait TargetIsa: Display + Send + Sync {
    fn name(&self) -> &'static str;

    fn triple(&self) -> &Triple;

    fn flags(&self) -> &Flags;

    fn isa_flags(&self) -> Vec<Value>;

    fn unsigned_add_overflow_condition(&self) -> IntCC;

    fn get_mach_backend(&self) -> &dyn MachBackend;

    fn map_regalloc_reg_to_dwarf(
        &self, 
        _: Reg
    ) -> Result<u16, RegisterMappingError> { ... }

    fn create_systemv_cie(&self) -> Option<CommonInformationEntry> { ... }
}

Methods that are specialized to a target ISA.

Implies a Display trait that shows the shared flags, as well as any ISA-specific flags.

Required methods

fn name(&self) -> &'static str

Get the name of this ISA.

fn triple(&self) -> &Triple

Get the target triple that was used to make this trait object.

fn flags(&self) -> &Flags

Get the ISA-independent flags that were used to make this trait object.

fn isa_flags(&self) -> Vec<Value>

Get the ISA-dependent flag values that were used to make this trait object.

fn unsigned_add_overflow_condition(&self) -> IntCC

IntCC condition for Unsigned Addition Overflow (Carry).

fn get_mach_backend(&self) -> &dyn MachBackend

Get the new-style MachBackend.

Provided methods

fn map_regalloc_reg_to_dwarf(&self, _: Reg) -> Result<u16, RegisterMappingError>

Map a regalloc::Reg to its corresponding DWARF register.

fn create_systemv_cie(&self) -> Option<CommonInformationEntry>

Creates a new System V Common Information Entry for the ISA.

Returns None if the ISA does not support System V unwind information.

Implementations

impl<'a> dyn TargetIsa + 'a

Methods implemented for free for target ISA!

pub fn default_call_conv(&self) -> CallConv

Get the default calling convention of this target.

pub fn endianness(&self) -> Endianness

Get the endianness of this ISA.

pub fn code_section_alignment(&self) -> u64

Returns the code (text) section alignment for this ISA.

pub fn pointer_type(&self) -> Type

Get the pointer type of this ISA.

pub fn pointer_bits(&self) -> u8

Get the width of pointers on this ISA, in units of bits.

pub fn pointer_bytes(&self) -> u8

Get the width of pointers on this ISA, in units of bytes.

pub fn frontend_config(&self) -> TargetFrontendConfig

Get the information needed by frontends producing Cranelift IR.