Module cranelift_codegen::isa::x64::encoding

Contains the encoding machinery for the various x64 instruction formats.

Modules

evex

Encodes EVEX instructions. These instructions are those added by the AVX-512 extensions. The EVEX encoding requires a 4-byte prefix:

rex

Encodes instructions in the standard x86 encoding mode. This is called IA-32E mode in the Intel manuals but corresponds to the addition of the REX-prefix format (hence the name of this module) that allowed encoding instructions in both compatibility mode (32-bit instructions running on a 64-bit OS) and in 64-bit mode (using the full 64-bit address space).

vex

Encodes VEX instructions. These instructions are those added by the Advanced Vector Extensions (AVX).

Traits

ByteSink

The encoding formats in this module all require a way of placing bytes into a buffer.