Trait cranelift_codegen::machinst::abi::ABIBody

pub trait ABIBody {
    type I: VCodeInst;
    fn flags(&self) -> &Flags;

    fn liveins(&self) -> Set<RealReg>;

    fn liveouts(&self) -> Set<RealReg>;

    fn num_args(&self) -> usize;

    fn num_retvals(&self) -> usize;

    fn num_stackslots(&self) -> usize;

    fn gen_copy_arg_to_reg(
        &self, 
        idx: usize, 
        into_reg: Writable<Reg>
    ) -> Self::I;

    fn gen_copy_reg_to_retval(
        &self, 
        idx: usize, 
        from_reg: Writable<Reg>, 
        ext: ArgumentExtension
    ) -> Vec<Self::I>;

    fn gen_ret(&self) -> Self::I;

    fn gen_epilogue_placeholder(&self) -> Self::I;

    fn set_num_spillslots(&mut self, slots: usize);

    fn set_clobbered(&mut self, clobbered: Set<Writable<RealReg>>);

    fn stackslot_addr(
        &self, 
        slot: StackSlot, 
        offset: u32, 
        into_reg: Writable<Reg>
    ) -> Self::I;

    fn load_stackslot(
        &self, 
        slot: StackSlot, 
        offset: u32, 
        ty: Type, 
        into_reg: Writable<Reg>
    ) -> Self::I;

    fn store_stackslot(
        &self, 
        slot: StackSlot, 
        offset: u32, 
        ty: Type, 
        from_reg: Reg
    ) -> Self::I;

    fn load_spillslot(
        &self, 
        slot: SpillSlot, 
        ty: Type, 
        into_reg: Writable<Reg>
    ) -> Self::I;

    fn store_spillslot(
        &self, 
        slot: SpillSlot, 
        ty: Type, 
        from_reg: Reg
    ) -> Self::I;

    fn gen_prologue(&mut self) -> Vec<Self::I>;

    fn gen_epilogue(&self) -> Vec<Self::I>;

    fn frame_size(&self) -> u32;

    fn get_spillslot_size(&self, rc: RegClass, ty: Type) -> u32;

    fn gen_spill(
        &self, 
        to_slot: SpillSlot, 
        from_reg: RealReg, 
        ty: Type
    ) -> Self::I;

    fn gen_reload(
        &self, 
        to_reg: Writable<RealReg>, 
        from_slot: SpillSlot, 
        ty: Type
    ) -> Self::I;
}

Trait implemented by an object that tracks ABI-related state (e.g., stack layout) and can generate code while emitting the body of a function.

Associated Types

type I: VCodeInst

The instruction type for the ISA associated with this ABI.

Loading content...
Required methods

fn flags(&self) -> &Flags

Get the settings controlling this function's compilation.

fn liveins(&self) -> Set<RealReg>

Get the liveins of the function.

fn liveouts(&self) -> Set<RealReg>

Get the liveouts of the function.

fn num_args(&self) -> usize

Number of arguments.

fn num_retvals(&self) -> usize

Number of return values.

fn num_stackslots(&self) -> usize

Number of stack slots (not spill slots).

fn gen_copy_arg_to_reg(&self, idx: usize, into_reg: Writable<Reg>) -> Self::I

Generate an instruction which copies an argument to a destination register.

fn gen_copy_reg_to_retval(
&self,
idx: usize,
from_reg: Writable<Reg>,
ext: ArgumentExtension
) -> Vec<Self::I>

Generate an instruction which copies a source register to a return value slot.

fn gen_ret(&self) -> Self::I

Generate a return instruction.

fn gen_epilogue_placeholder(&self) -> Self::I

Generate an epilogue placeholder. The returned instruction should return true from is_epilogue_placeholder(); this is used to indicate to the lowering driver when the epilogue should be inserted.

fn set_num_spillslots(&mut self, slots: usize)

Update with the number of spillslots, post-regalloc.

fn set_clobbered(&mut self, clobbered: Set<Writable<RealReg>>)

Update with the clobbered registers, post-regalloc.

fn stackslot_addr(
&self,
slot: StackSlot,
offset: u32,
into_reg: Writable<Reg>
) -> Self::I

Get the address of a stackslot.

fn load_stackslot(
&self,
slot: StackSlot,
offset: u32,
ty: Type,
into_reg: Writable<Reg>
) -> Self::I

Load from a stackslot.

fn store_stackslot(
&self,
slot: StackSlot,
offset: u32,
ty: Type,
from_reg: Reg
) -> Self::I

Store to a stackslot.

fn load_spillslot(
&self,
slot: SpillSlot,
ty: Type,
into_reg: Writable<Reg>
) -> Self::I

Load from a spillslot.

fn store_spillslot(&self, slot: SpillSlot, ty: Type, from_reg: Reg) -> Self::I

Store to a spillslot.

fn gen_prologue(&mut self) -> Vec<Self::I>

Generate a prologue, post-regalloc. This should include any stack frame or other setup necessary to use the other methods (load_arg, store_retval, and spillslot accesses.) self is mutable so that we can store information in it which will be useful when creating the epilogue.

fn gen_epilogue(&self) -> Vec<Self::I>

Generate an epilogue, post-regalloc. Note that this must generate the actual return instruction (rather than emitting this in the lowering logic), because the epilogue code comes before the return and the two are likely closely related.

fn frame_size(&self) -> u32

Returns the full frame size for the given function, after prologue emission has run. This comprises the spill slots and stack-storage slots (but not storage for clobbered callee-save registers, arguments pushed at callsites within this function, or other ephemeral pushes). This is used for ABI variants where the client generates prologue/epilogue code, as in Baldrdash (SpiderMonkey integration).

fn get_spillslot_size(&self, rc: RegClass, ty: Type) -> u32

Get the spill-slot size.

fn gen_spill(&self, to_slot: SpillSlot, from_reg: RealReg, ty: Type) -> Self::I

Generate a spill.

fn gen_reload(
&self,
to_reg: Writable<RealReg>,
from_slot: SpillSlot,
ty: Type
) -> Self::I

Generate a reload (fill).

Loading content...