Struct cranelift_codegen::machinst::abi_impl::ABICallerImpl

pub struct ABICallerImpl<M: ABIMachineSpec> { /* private fields */ }

ABI object for a callsite.

Implementations

impl<M: ABIMachineSpec> ABICallerImpl<M>

pub fn from_func(
sig: &Signature,
extname: &ExternalName,
dist: RelocDistance,
caller_conv: CallConv
) -> CodegenResult<ABICallerImpl<M>>

Create a callsite ABI object for a call directly to the specified function.

pub fn from_ptr(
sig: &Signature,
ptr: Reg,
opcode: Opcode,
caller_conv: CallConv
) -> CodegenResult<ABICallerImpl<M>>

Create a callsite ABI object for a call to a function pointer with the given signature.