Module cranelift_codegen::machinst::lower

This module implements lowering (instruction selection) from Cranelift IR to machine instructions with virtual registers. This is almost the final machine code, except for register allocation.

Structs

Lower
Machine-independent lowering driver / machine-instruction container. Maintains a correspondence from original Inst to MachInsts.

NonRegInput
A representation of all of the ways in which a value is available, aside from as a direct register.

Enums

RelocDistance
Notion of “relocation distance”. This gives an estimate of how far away a symbol will be from a reference.

Traits

LowerBackend
A machine backend.

LowerCtx
A context that machine-specific lowering code can use to emit lowered instructions. This is the view of the machine-independent per-function lowering context that is seen by the machine backend.