Trait cranelift_codegen::machinst::lower::LowerCtx

pub trait LowerCtx {
    type I;
    fn data(&self, ir_inst: Inst) -> &InstructionData;

    fn ty(&self, ir_inst: Inst) -> Type;

    fn abi(&mut self) -> &dyn ABIBody<I = Self::I>;

    fn emit(&mut self, mach_inst: Self::I);

    fn merged(&mut self, from_inst: Inst);

    fn input_inst(&self, ir_inst: Inst, idx: usize) -> Option<(Inst, usize)>;

    fn value_to_writable_reg(&self, val: Value) -> Writable<Reg>;

    fn value_to_reg(&self, val: Value) -> Reg;

    fn input(&self, ir_inst: Inst, idx: usize) -> Reg;

    fn output(&self, ir_inst: Inst, idx: usize) -> Writable<Reg>;

    fn num_inputs(&self, ir_inst: Inst) -> usize;

    fn num_outputs(&self, ir_inst: Inst) -> usize;

    fn input_ty(&self, ir_inst: Inst, idx: usize) -> Type;

    fn output_ty(&self, ir_inst: Inst, idx: usize) -> Type;

    fn tmp(&mut self, rc: RegClass, ty: Type) -> Writable<Reg>;

    fn num_bb_params(&self, bb: Block) -> usize;

    fn bb_param(&self, bb: Block, idx: usize) -> Reg;

    fn retval(&self, idx: usize) -> Writable<Reg>;

    fn call_target<'b>(
        &'b self, 
        ir_inst: Inst
    ) -> Option<(&'b ExternalName, RelocDistance)>;

    fn call_sig<'b>(&'b self, ir_inst: Inst) -> Option<&'b Signature>;

    fn symbol_value<'b>(
        &'b self, 
        ir_inst: Inst
    ) -> Option<(&'b ExternalName, RelocDistance, i64)>;

    fn memflags(&self, ir_inst: Inst) -> Option<MemFlags>;

    fn srcloc(&self, ir_inst: Inst) -> SourceLoc;
}

A context that machine-specific lowering code can use to emit lowered instructions. This is the view of the machine-independent per-function lowering context that is seen by the machine backend.

Associated Types

type I

The instruction type for which this lowering framework is instantiated.

Loading content...
Required methods

fn data(&self, ir_inst: Inst) -> &InstructionData

Get the instdata for a given IR instruction.

fn ty(&self, ir_inst: Inst) -> Type

Get the controlling type for a polymorphic IR instruction.

fn abi(&mut self) -> &dyn ABIBody<I = Self::I>

Get the ABIBody.

fn emit(&mut self, mach_inst: Self::I)

Emit a machine instruction.

fn merged(&mut self, from_inst: Inst)

Indicate that an IR instruction has been merged, and so one of its uses is gone (replaced by uses of the instruction's inputs). This helps the lowering algorithm to perform on-the-fly DCE, skipping over unused instructions (such as immediates incorporated directly).

fn input_inst(&self, ir_inst: Inst, idx: usize) -> Option<(Inst, usize)>

Get the producing instruction, if any, and output number, for the idxth input to the given IR instruction

fn value_to_writable_reg(&self, val: Value) -> Writable<Reg>

Map a Value to its associated writable (probably virtual) Reg.

fn value_to_reg(&self, val: Value) -> Reg

Map a Value to its associated (probably virtual) Reg.

fn input(&self, ir_inst: Inst, idx: usize) -> Reg

Get the idxth input to the given IR instruction as a virtual register.

fn output(&self, ir_inst: Inst, idx: usize) -> Writable<Reg>

Get the idxth output of the given IR instruction as a virtual register.

fn num_inputs(&self, ir_inst: Inst) -> usize

Get the number of inputs to the given IR instruction.

fn num_outputs(&self, ir_inst: Inst) -> usize

Get the number of outputs to the given IR instruction.

fn input_ty(&self, ir_inst: Inst, idx: usize) -> Type

Get the type for an instruction's input.

fn output_ty(&self, ir_inst: Inst, idx: usize) -> Type

Get the type for an instruction's output.

fn tmp(&mut self, rc: RegClass, ty: Type) -> Writable<Reg>

Get a new temp.

fn num_bb_params(&self, bb: Block) -> usize

Get the number of block params.

fn bb_param(&self, bb: Block, idx: usize) -> Reg

Get the register for a block param.

fn retval(&self, idx: usize) -> Writable<Reg>

Get the register for a return value.

fn call_target<'b>(
&'b self,
ir_inst: Inst
) -> Option<(&'b ExternalName, RelocDistance)>

Get the target for a call instruction, as an ExternalName. Returns a tuple providing this name and the "relocation distance", i.e., whether the backend can assume the target will be "nearby" (within some small offset) or an arbitrary address. (This comes from the colocated bit in the CLIF.)

fn call_sig<'b>(&'b self, ir_inst: Inst) -> Option<&'b Signature>

Get the signature for a call or call-indirect instruction.

fn symbol_value<'b>(
&'b self,
ir_inst: Inst
) -> Option<(&'b ExternalName, RelocDistance, i64)>

Get the symbol name, relocation distance estimate, and offset for a symbol_value instruction.

fn memflags(&self, ir_inst: Inst) -> Option<MemFlags>

Returns the memory flags of a given memory access.

fn srcloc(&self, ir_inst: Inst) -> SourceLoc

Get the source location for a given instruction.

Loading content...