Trait cranelift_codegen::machinst::MachInstEmitState

pub trait MachInstEmitState<I: MachInst>: Default + Clone + Debug {
    fn new(abi: &dyn ABICallee<I = I>) -> Self;

    fn pre_safepoint(&mut self, _stack_map: StackMap) { ... }

    fn pre_sourceloc(&mut self, _srcloc: SourceLoc) { ... }
}

A trait describing the emission state carried between MachInsts when emitting a function body.

Required methods

fn new(abi: &dyn ABICallee<I = I>) -> Self

Create a new emission state given the ABI object.

Loading content...
Provided methods

fn pre_safepoint(&mut self, _stack_map: StackMap)

Update the emission state before emitting an instruction that is a safepoint.

fn pre_sourceloc(&mut self, _srcloc: SourceLoc)

Update the emission state to indicate instructions are associated with a particular SourceLoc.

Loading content...