Module cranelift_codegen::machinst::valueregs

Data structure for tracking the (possibly multiple) registers that hold one SSA Value.

Structs

ValueRegs
Location at which a Value is stored in register(s): the value is located in one or more registers, depending on its width. A value may be stored in more than one register if the machine has no registers wide enough otherwise: for example, on a 32-bit architecture, we may store I64 values in two registers, and I128 values in four.

Traits

InvalidSentinel
A type with an “invalid” sentinel value.