Struct cranelift_codegen::machinst::vcode::VCode

pub struct VCode<I: VCodeInst> {
    pub fallthrough_return_block: Option<BlockIndex>,
    // some fields omitted
}

A function in "VCode" (virtualized-register code) form, after lowering. This is essentially a standard CFG of basic blocks, where each basic block consists of lowered instructions produced by the machine-specific backend.

Fields
fallthrough_return_block: Option<BlockIndex>

The block targeted by fallthrough_returns, if there's one.

Implementations

impl<I: VCodeInst> VCode<I>

pub fn flags(&self) -> &Flags

Returns the flags controlling this function's compilation.

pub fn vreg_type(&self, vreg: VirtualReg) -> Type

Get the IR-level type of a VReg.

pub fn entry(&self) -> BlockIndex

Get the entry block.

pub fn num_blocks(&self) -> usize

Get the number of blocks. Block indices will be in the range 0 .. (self.num_blocks() - 1).

pub fn frame_size(&self) -> u32

Stack frame size for the full function's body.

pub fn succs(&self, block: BlockIndex) -> &[BlockIx]

Get the successors for a block.

pub fn replace_insns_from_regalloc(&mut self, result: RegAllocResult<Self>)

Take the results of register allocation, with a sequence of instructions including spliced fill/reload/move instructions, and replace the VCode with them.

pub fn remove_redundant_branches(&mut self)

Removes redundant branches, rewriting targets to point directly to the ultimate block at the end of a chain of trivial one-target jumps.

pub fn finalize_branches(&mut self) where
I: MachInstEmit<MachSectionSize>,

Mutate branch instructions to (i) lower two-way condbrs to one-way, depending on fallthrough; and (ii) use concrete offsets.

pub fn emit(&self) -> MachSections where
I: MachInstEmit<MachSection>,

Emit the instructions to a list of sections.

pub fn bindex_to_bb(&self, block: BlockIndex) -> Option<Block>

Get the IR block for a BlockIndex, if one exists.