Struct cranelift_codegen::settings::Flags

pub struct Flags { /* fields omitted */ }

Flags group shared.

Methods

impl Flags

pub fn new(builder: Builder) -> Self

Create flags shared settings group.

impl Flags

User-defined settings.

pub fn predicate_view(&self) -> PredicateView

Get a view of the boolean predicates.

pub fn opt_level(&self) -> OptLevel

Optimization level:

none: Minimise compile time by disabling most optimizations.

speed: Generate the fastest possible code

speed_and_size: like "speed", but also perform transformations aimed at reducing code size.

pub fn libcall_call_conv(&self) -> LibcallCallConv

Defines the calling convention to use for LibCalls call expansion, since it may be different from the ISA default calling convention.

The default value is to use the same calling convention as the ISA default calling convention.

This list should be kept in sync with the list of calling conventions available in isa/call_conv.rs.

pub fn baldrdash_prologue_words(&self) -> u8

Number of pointer-sized words pushed by the baldrdash prologue.

Functions with the baldrdash calling convention don't generate their own prologue and epilogue. They depend on externally generated code that pushes a fixed number of words in the prologue and restores them in the epilogue.

This setting configures the number of pointer-sized words pushed on the stack when the Cranelift-generated code is entered. This includes the pushed return address on x86.

pub fn probestack_size_log2(&self) -> u8

The log2 of the size of the stack guard region.

Stack frames larger than this size will have stack overflow checked by calling the probestack function.

The default is 12, which translates to a size of 4096.

pub fn enable_verifier(&self) -> bool

Run the Cranelift IR verifier at strategic times during compilation.

This makes compilation slower but catches many bugs. The verifier is always enabled by default, which is useful during development.

pub fn is_pic(&self) -> bool

Enable Position-Independent Code generation

pub fn colocated_libcalls(&self) -> bool

Use colocated libcalls.

Generate code that assumes that libcalls can be declared "colocated", meaning they will be defined along with the current function, such that they can use more efficient addressing.

pub fn avoid_div_traps(&self) -> bool

Generate explicit checks around native division instructions to avoid their trapping.

This is primarily used by SpiderMonkey which doesn't install a signal handler for SIGFPE, but expects a SIGILL trap for division by zero.

On ISAs like ARM where the native division instructions don't trap, this setting has no effect - explicit checks are always inserted.

pub fn enable_float(&self) -> bool

Enable the use of floating-point instructions

Disabling use of floating-point instructions is not yet implemented.

pub fn enable_nan_canonicalization(&self) -> bool

Enable NaN canonicalization

This replaces NaNs with a single canonical value, for users requiring entirely deterministic WebAssembly computation. This is not required by the WebAssembly spec, so it is not enabled by default.

pub fn enable_pinned_reg(&self) -> bool

Enable the use of the pinned register.

This register is excluded from register allocation, and is completely under the control of the end-user. It is possible to read it via the get_pinned_reg instruction, and to set it with the set_pinned_reg instruction.

pub fn use_pinned_reg_as_heap_base(&self) -> bool

Use the pinned register as the heap base.

Enabling this requires the enable_pinned_reg setting to be set to true. It enables a custom legalization of the heap_addr instruction so it will use the pinned register as the heap base, instead of fetching it from a global value.

Warning! Enabling this means that the pinned register must be maintained to contain the heap base address at all times, during the lifetime of a function. Using the pinned register for other purposes when this is set is very likely to cause crashes.

pub fn enable_simd(&self) -> bool

Enable the use of SIMD instructions.

pub fn enable_atomics(&self) -> bool

Enable the use of atomic instructions

pub fn enable_safepoints(&self) -> bool

Enable safepoint instruction insertions.

This will allow the emit_stackmaps() function to insert the safepoint instruction on top of calls and interrupt traps in order to display the live reference values at that point in the program.

pub fn allones_funcaddrs(&self) -> bool

Emit not-yet-relocated function addresses as all-ones bit patterns.

pub fn probestack_enabled(&self) -> bool

Enable the use of stack probes, for calling conventions which support this functionality.

pub fn probestack_func_adjusts_sp(&self) -> bool

Set this to true of the stack probe function modifies the stack pointer itself.

pub fn jump_tables_enabled(&self) -> bool

Enable the use of jump tables in generated machine code.