Struct cranelift_codegen::MachTextSectionBuilder

pub struct MachTextSectionBuilder<I: VCodeInst> { /* private fields */ }

Implementation of the TextSectionBuilder trait backed by MachBuffer.

Note that MachBuffer was primarily written for intra-function references of jumps between basic blocks, but it’s also quite usable for entire text sections and resolving references between functions themselves. This builder interprets “blocks” as labeled functions for the purposes of resolving labels internally in the buffer.

Implementations

impl<I: VCodeInst> MachTextSectionBuilder<I>

pub fn new(num_funcs: usize) -> MachTextSectionBuilder<I>

Creates a new text section builder which will have num_funcs functions pushed into it.