Module cranelift_codegen::timing

Pass timing.

This modules provides facilities for timing the execution of individual compilation passes.

Structs

DefaultProfiler

The default profiler. You can get the results using take_current.

PassTimes

Accumulated timing for all passes.

Enums

Pass

A single profiled pass.

Constants

NUM_PASSES

The amount of profiled passes.

Traits

Profiler

A profiler.

Functions

canonicalize_nans

Canonicalization of NaNs

compile

Compilation passes

domtree

Dominator tree

egraph

Egraph based optimizations

flowgraph

Control flow graph

gvn

Global value numbering

layout_renumber

Layout full renumbering

licm

Loop invariant code motion

loop_analysis

Loop analysis

parse_text

Parsing textual Cranelift IR

preopt

Pre-legalization rewriting

process_file

Processing test file

regalloc

Register allocation

regalloc_checker

Register allocation symbolic verification

remove_constant_phis

Remove constant phi-nodes

set_thread_profiler

Set the profiler for the current thread.

store_incremental_cache

Store in incremental cache

take_current

Take the current accumulated pass timings and reset the timings for the current thread.

try_incremental_cache

Try loading from incremental cache

unreachable_code

Remove unreachable blocks

vcode_emit

VCode emission

vcode_emit_finish

VCode emission finalization

vcode_lower

VCode lowering

verifier

Verify Cranelift IR

wasm_translate_function

Translate WASM function

wasm_translate_module

Translate WASM module