Module cranelift_codegen::write

Converting Cranelift IR to text.

The write module provides the write_function function which converts an IR Function to an equivalent textual form. This textual form can be read back by the cranelift-reader crate.

Structs

PlainWriter

A PlainWriter that doesn’t decorate the function.

Traits

FuncWriter

A FuncWriter used to decorate functions during printing.

Functions

decorate_function

Writes func to w as text. write_function_plain is passed as ‘closure’ to print instructions as text. pretty_function_error is passed as ‘closure’ to add error decoration.

write_block_header

Write out the basic block header, outdented:

write_function

Write func to w as equivalent text. Use isa to emit ISA-dependent annotations.

write_operands

Write the operands of inst to w with a prepended space.