Trait cranelift_codegen::write::FuncWriter

pub trait FuncWriter {
    fn write_block_header(
        &mut self, 
        w: &mut dyn Write, 
        func: &Function, 
        isa: Option<&dyn TargetIsa>, 
        block: Block, 
        indent: usize
    ) -> Result;

    fn write_instruction(
        &mut self, 
        w: &mut dyn Write, 
        func: &Function, 
        aliases: &SecondaryMap<Value, Vec<Value>>, 
        isa: Option<&dyn TargetIsa>, 
        inst: Inst, 
        indent: usize
    ) -> Result;

    fn write_preamble(
        &mut self, 
        w: &mut dyn Write, 
        func: &Function, 
        regs: Option<&RegInfo>
    ) -> Result<bool, Error> { ... }

    fn super_preamble(
        &mut self, 
        w: &mut dyn Write, 
        func: &Function, 
        regs: Option<&RegInfo>
    ) -> Result<bool, Error> { ... }

    fn write_entity_definition(
        &mut self, 
        w: &mut dyn Write, 
        func: &Function, 
        entity: AnyEntity, 
        value: &dyn Display
    ) -> Result { ... }

    fn super_entity_definition(
        &mut self, 
        w: &mut dyn Write, 
        func: &Function, 
        entity: AnyEntity, 
        value: &dyn Display
    ) -> Result { ... }
}

A FuncWriter used to decorate functions during printing.

Required methods

fn write_block_header(
&mut self,
w: &mut dyn Write,
func: &Function,
isa: Option<&dyn TargetIsa>,
block: Block,
indent: usize
) -> Result

Write the basic block header for the current function.

fn write_instruction(
&mut self,
w: &mut dyn Write,
func: &Function,
aliases: &SecondaryMap<Value, Vec<Value>>,
isa: Option<&dyn TargetIsa>,
inst: Inst,
indent: usize
) -> Result

Write the given inst to w.

Provided methods

fn write_preamble(
&mut self,
w: &mut dyn Write,
func: &Function,
regs: Option<&RegInfo>
) -> Result<bool, Error>

Write the preamble to w. By default, this uses write_entity_definition.

fn super_preamble(
&mut self,
w: &mut dyn Write,
func: &Function,
regs: Option<&RegInfo>
) -> Result<bool, Error>

Default impl of write_preamble

fn write_entity_definition(
&mut self,
w: &mut dyn Write,
func: &Function,
entity: AnyEntity,
value: &dyn Display
) -> Result

Write an entity definition defined in the preamble to w.

fn super_entity_definition(
&mut self,
w: &mut dyn Write,
func: &Function,
entity: AnyEntity,
value: &dyn Display
) -> Result

Default impl of write_entity_definition